+++ mcodes/PXO_to_MTEX_to_MTEX2GMSH_to_ABAQUS.m
@@ -1,0 +1,14 @@
+
+
+
+G = gmshGeo(GRAINS);
+
+mesh(G,'default.msh')
+
+mesh(G,'constant_elmtSize.msh','ElementSize',50)
+
+mesh(G,'Quad.msh','ElementType','Quad');
+
+mesh(G,'quad_abaqus.inp','ElementType','Quad');
+
+]
