+++ mcodes/Marketing_Progress.m
@@ -1,26 +1,0 @@
-function Marketing_Progress()
-
-global RG LI FB
-
-RG.date        = {'18Aug2020'};
-RG.updates     = [7];
-RG.reads       = [160];
-RG.follow      = [8];
-RG.updates     = [7];
-RG.comments    = [0];
-RG.publication = [0];
-
-LI.date        = {'18Aug2020'};
-LI.updates     = [1];
-LI.views       = [1640];
-LI.like        = [33];
-LI.celebrate   = [1];
-LI.insightful  = [1];
-LI.publication = [0];
-
-FB.date        = {'18Aug2020'};
-FB.updates     = [2];
-FB.comments    = [11];
-FB.reactions   = [132];
-
-end
