+++ mcodes/results/colormatrix.txt
@@ -1,32 +1,32 @@
-0.18943	0.25517	0.34558
-0.43545	0.73993	0.81274
-0.91241	0.40561	0.004961
-0.19012	0.53923	0.23773
-0.70902	0.15875	0.031866
-0.76364	0.96961	0.67694
-0.27215	0.56556	0.85161
-0.29507	0.13014	0.76941
-0.96122	0.531	0.57645
-0.2134	0.23743	0.2688
-0.38059	0.86002	0.91612
-0.66773	0.91989	0.83665
-0.37416	0.58504	0.93032
-0.68313	0.34099	0.8864
-0.97532	0.99418	0.76456
-0.66892	0.68112	0.6079
-0.69743	0.17886	0.74767
-0.2475	0.23584	0.2982
-0.87253	0.7364	0.72126
-0.77657	0.55546	0.90146
-0.14008	0.039626	0.044069
-0.25278	0.55715	0.34249
-0.55708	0.97523	0.2384
-0.50985	0.70883	0.070693
-0.97601	0.96641	0.44843
-0.77991	0.89962	0.070638
-0.24295	0.58017	0.90404
-0.08868	0.71334	0.45107
-0.80345	0.84369	0.52784
-0.70319	0.68829	0.97139
-0.92565	0.16247	0.79868
-0.40248	0.1404	0.52446
+0.83706	0.10779	0.57768
+0.24277	0.8689	0.34238
+0.50573	0.0032652	0.13547
+0.46839	0.51859	0.50839
+0.8049	0.55082	0.64626
+0.48547	0.11595	0.96752
+0.047994	0.20004	0.76195
+0.29891	0.86222	0.183
+0.72163	0.11697	0.23966
+0.54076	0.8247	0.95701
+0.71301	0.19309	0.758
+0.89043	0.99555	0.19292
+0.036948	0.47141	0.70195
+0.16157	0.54727	0.83347
+0.20618	0.082274	0.43625
+0.54819	0.051293	0.0094767
+0.6875	0.20742	0.66811
+0.34233	0.16079	0.069735
+0.8698	0.45106	0.2378
+0.8792	0.62995	0.097031
+0.19012	0.72633	0.87047
+0.53183	0.45605	0.83404
+0.98953	0.37693	0.23516
+0.057573	0.064415	0.75438
+0.083431	0.53834	0.50836
+0.41763	0.45104	0.42116
+0.85461	0.40236	0.21872
+0.39859	0.073837	0.22166
+0.0099489	0.19382	0.090802
+0.64554	0.60843	0.24882
+0.2938	0.91508	0.58965
+0.14993	0.63152	0.88221
