+++ mcodes/simparameters/initialfinalmcs.txt
@@ -1,2 +1,2 @@
 1
-1000
+2000
