### 2021-02-08 b79aee0
updates for test case images
