+++ mcodes/PXO_to_MTEX_to_MTEX2GMSH_to_ABAQUS.m
@@ -10,5 +10,3 @@ mesh(G,'constant_elmtSize.msh','ElementSize',50)
 mesh(G,'Quad.msh','ElementType','Quad');
 
 mesh(G,'quad_abaqus.inp','ElementType','Quad');
-
-]
