+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using49_VERTICAL_intercepts.txt
@@ -1,0 +1,4 @@
+5.1213
+6.7498
+7.711
+8.3799
