+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using15_VERTICAL_intercepts.txt
@@ -1,0 +1,4 @@
+5.0539
+6.6288
+8.0289
+8.3453
