+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_3_incr =7.8963.txt
@@ -1,0 +1,1 @@
+7.8963
