+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using24_HORIZONTAL_intercepts.txt
@@ -1,0 +1,4 @@
+4.973
+6.7166
+7.1445
+7.537
