+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using49_HORIZONTAL_intercepts.txt
@@ -1,0 +1,4 @@
+4.9843
+6.6828
+7.3292
+7.634
