+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using15_HORIZONTAL_intercepts.txt
@@ -1,0 +1,4 @@
+4.7676
+6.4132
+6.9705
+7.4472
