+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_2_incr =7.9813.txt
@@ -1,0 +1,1 @@
+7.9813
