+++ mcodes/results/datafiles/graindata/grainsizes/A_A_AGSvsMCS_M_intincr_1_2_3_and_4 =7.9538.txt
@@ -1,0 +1,1 @@
+7.9538
