### 2021-03-21 #3018 @Kevin-Mattheus-Moerman
@SunilAnandatheertha apologies for the delay in processing your submission. JOSS is experiencing delays since both editors and reviewers are experiencing additional time constraints and difficulties due to the COVID19 pandemic.
I am picking this up now and will hopefully have found and assigned reviewers this. Thanks for your patience.
