### 2021-04-16 #3018 @Kevin-Mattheus-Moerman
@ksyang2013, @jan-janssen, @jic198, @hitliaomq, @DavidMercier, @czambaldi , @ralfHielscher, @frankNiessen, @millerta, @ketch, @yangbai90, would you be interested in reviewing this submission  for [JOSS](https://joss.theoj.org/)? The [review process](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process) focuses on [the software](https://github.com/SunilAnandatheertha/PXO) and this [short paper](https://github.com/openjournals/joss-papers/blob/joss.03018/joss.03018/10.21105.joss.03018.pdf). If you are interested you can comment here. Thanks!

Note, given the global COVID19 pandemic we can be very flexible in terms of review times.

Note, this is a so-called "Pre-review issue" where reviewers/editors are assigned. Once I've attracted sufficient reviewers I'll trigger a "Review issue" where the review will take place.
