### 2021-04-17 #3018 @hitliaomq
Dear Kevin,

Thank you for your invitation.

But I have no time recently, I must focus on my Ph.D thesis. In the future (after september), I'm glad to review papers or codes.

Thank you again.

Best,
Mingqing

On 4/16/2021 16:11，Kevin Mattheus ***@***.***> wrote：

@ksyang2013, @jan-janssen, @jic198, @hitliaomq, @DavidMercier, @czambaldi , @ralfHielscher, @frankNiessen, @millerta, @ketch, @yangbai90, would you be interested in reviewing this submission for JOSS? The review process focuses on the software and this short paper. If you are interested you can comment here. Thanks!

Note, given the global COVID19 pandemic we can be very flexible in terms of review times.

Note, this is a so-called "Pre-review issue" where reviewers/editors are assigned. Once I've attracted sufficient reviewers I'll trigger a "Review issue" where the review will take place.
