### 2021-04-17 #3018 @ksyang2013
Dear Kevin,

I am willing to review this codebase. Thanks!

Best,
Kesong

--------------------------------------
Kesong YANG, Ph.D.

Associate Professor,
Department of NanoEngineering,
Program of Materials Science and Engineering,
Center for Memory and Recording Research (CMRR),
University of California San Diego
9500 Gilman Drive, Mail Code 0448
La Jolla, CA 92093-0448, USA

Office: SME 244H
Tel: (858) 534-2514
E-mail: ***@***.***
http://materials.ucsd.edu <http://materials.ucsd.edu/>

> On Apr 16, 2021, at 1:11 AM, Kevin Mattheus Moerman ***@***.***> wrote:
>
>
> @ksyang2013 <https://urldefense.com/v3/__https://github.com/ksyang2013__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVH1xhKLqc$>, @jan-janssen <https://urldefense.com/v3/__https://github.com/jan-janssen__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVH_LtZTcU$>, @jic198 <https://urldefense.com/v3/__https://github.com/jic198__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHP2B0-cw$>, @hitliaomq <https://urldefense.com/v3/__https://github.com/hitliaomq__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVH_OXRxrU$>, @DavidMercier <https://urldefense.com/v3/__https://github.com/DavidMercier__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVH8BiNNac$>, @czambaldi <https://urldefense.com/v3/__https://github.com/czambaldi__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHyFXCXt8$> , @ralfHielscher <https://urldefense.com/v3/__https://github.com/ralfHielscher__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHDdmi5Gs$>, @frankNiessen <https://urldefense.com/v3/__https://github.com/frankNiessen__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHhTAjkmg$>, @millerta <https://urldefense.com/v3/__https://github.com/millerta__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHHH_LyE8$>, @ketch <https://urldefense.com/v3/__https://github.com/ketch__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVH9VUgy04$>, @yangbai90 <https://urldefense.com/v3/__https://github.com/yangbai90__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVH59yxDTs$>, would you be interested in reviewing this submission for JOSS <https://urldefense.com/v3/__https://joss.theoj.org/__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHN4CYc1A$>? The review process <https://urldefense.com/v3/__https://joss.readthedocs.io/en/latest/submitting.html*the-review-process__;Iw!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHsVQPb7U$> focuses on the software <https://urldefense.com/v3/__https://github.com/SunilAnandatheertha/PXO__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHTITv6yA$> and this short paper <https://urldefense.com/v3/__https://github.com/openjournals/joss-papers/blob/joss.03018/joss.03018/10.21105.joss.03018.pdf__;!!Mih3wA!VpGlliBL6FEw00uWHPgxJ6rVU8dp9t1CmTecS1aSzveVZ7fRbbgiGZVHPjudqRU$>. If you are interested you can comment here. Thanks!
>
> Note, given the global COVID19 pandemic we can be very flexible in terms of review times.
>
> Note, this is a so-called "Pre-review issue" where reviewers/editors are assigned. Once I've attracted sufficient reviewers I'll trigger a "Review issue" where the review will take place.
>
