### 2021-04-29 #3190 @littlelazy6
Hi

I propose to revert next weekend, no later than 14th May. Hope this suits.

Regards
Yuhui

From: YANG Research Group@UC San Diego ***@***.***>
Sent: 2021年4月28日 18:41
Cc: Yuhui Tu ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [REVIEW]: PXO (Poly-XTAL Operations): Free MATLAB codebase to generate and analyse complex 2D grain structures (#3190)

Will return the comments within two weeks. Thanks!

--------------------------------------
Kesong YANG, Ph.D.

Associate Professor,
Department of NanoEngineering,
Program of Materials Science and Engineering,
Center for Memory and Recording Research (CMRR),
University of California San Diego
9500 Gilman Drive, Mail Code 0448
La Jolla, CA 92093-0448, USA

Office: SME 244H
Tel: (858) 534-2514
E-mail: ***@***.***
http://materials.ucsd.edu <http://materials.ucsd.edu/>

> On Apr 28, 2021, at 8:38 AM, Kevin Mattheus Moerman ***@***.***> wrote:
>
>
> @yangbai90 <https://urldefense.com/v3/__https://github.com/yangbai90__;!!Mih3wA!Twc8jCbmNKmR0OJthyOaAw0sAqUeesKDtF9jqKfdufVzI6VNvlv09dheLqHOPJQ$>, @ksyang2013 <https://urldefense.com/v3/__https://github.com/ksyang2013__;!!Mih3wA!Twc8jCbmNKmR0OJthyOaAw0sAqUeesKDtF9jqKfdufVzI6VNvlv09dhe8N_4YnQ$>, @littlelazy6 <https://urldefense.com/v3/__https://github.com/littlelazy6__;!!Mih3wA!Twc8jCbmNKmR0OJthyOaAw0sAqUeesKDtF9jqKfdufVzI6VNvlv09dheCZgtjS0$> this is where the review takes place. Could you provide an indication as to the time you need to provide your initial review? FYI we can be flexible and it is okay if you need to take quite a bit of time. Thanks again for your help!!!!
>
> ―
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub <https://urldefense.com/v3/__https://github.com/openjournals/joss-reviews/issues/3190*issuecomment-828557919__;Iw!!Mih3wA!Twc8jCbmNKmR0OJthyOaAw0sAqUeesKDtF9jqKfdufVzI6VNvlv09dheG6d7B4M$>, or unsubscribe <https://urldefense.com/v3/__https://github.com/notifications/unsubscribe-auth/AEHTG3HZR7XSCSTVSAUZTY3TLATYVANCNFSM43EGUGCA__;!!Mih3wA!Twc8jCbmNKmR0OJthyOaAw0sAqUeesKDtF9jqKfdufVzI6VNvlv09dheIUpBBxg$>.
>

―
