### 2021-05-21 #3190 @littlelazy6
Installation & general test case manual:
Please give definition for CNT, SLSP and SLSPC at the first instance. I firstly saw them in Limitations.
Where is the STEP 5 in the installation instruction?

Functionality documentation:
As a starting point, test case 01 is like a basic tutorial and I am referring to https://github.com/SunilAnandatheertha/PXO/wiki/Test-case---01
It could assist an easier and quicker start given that more details were added into this documentation.
1.	Voronoi tessellation (VT) model generating method. The first section highlights the regions A to E, to explain the generated resulting files. Could you add in a section explaining where to change the key VT setting in source code, such as seed (number of grain), aspect ratio (grain morphology), model size, mesh density? I think this is one core capability of PXO.
2.	Could you also briefly explain where in the source code to modify the kernel parameters for Monte Carlo MC simulation? It would be great to indicate what specific material you are simulating for the test case.
