### 2021-05-21 #3190 @littlelazy6
State of the field: Functions comparing to other software. This part seems to be missing in the current paper draft.
Based on my knowledge, there are several open-access packages available for MC based grain growth (GG) simulation solver, such as MCPM. VT generating tools include Python, NEPER, DREAM3D. Could you please add in few lines showing what distinguish your developed tool from other similar software? For e.g., do you use improved MC theory, grain size measurement or VT algorithm? Linkage to MTEX and ABAQUS could be a strong novelty in this work and is worth emphasizing when compared to other tools.

Generally speaking, this is a handy and useful software from my perspective. Below is just a suggestion for future development. PXO currently provides plenty of output channels to EBSD, FE mesh data. It would be beneficial to offer more choice for inputs source. If the user generates polycrystalline map from 3rd party tool and convert the nodal, phase, and orientation info into a PXO supported format, such structure can then be imported into PXO and enjoy all functions. This capability may broaden the popularity of this tool.
