+++ mcodes/results/datafiles/statematrices/initialsmatrix.txt
@@ -1,50 +1,0 @@
-22	24	25	18	26	19	13	13	32	12	15	18	18	22	8	19	4	30	23	4	22	4	28	3	1	13	15	2	23	8	16	30	25	12	16	19	24	14	2	21	6	29	2	16	27	4	20	30	23	25
-25	5	27	22	16	10	21	18	25	29	14	3	22	17	6	1	27	14	1	10	30	32	23	10	19	27	1	3	13	18	29	5	26	13	26	11	24	27	7	6	13	4	31	1	22	15	21	12	16	7
-2	13	28	13	14	21	30	25	13	28	25	4	26	6	8	14	13	14	15	5	29	17	13	7	26	11	4	28	12	26	1	13	16	20	16	23	9	2	12	12	1	17	18	15	9	5	24	17	16	12
-32	11	5	25	20	17	14	31	14	17	14	2	9	23	29	17	32	24	8	15	13	22	30	17	32	17	4	5	5	5	22	19	29	31	32	30	14	5	21	19	13	32	18	7	5	22	30	26	8	2
-3	29	29	29	13	27	17	5	23	2	22	7	9	18	24	27	30	20	26	21	20	24	28	21	32	20	13	22	1	18	21	12	3	8	29	3	19	19	19	26	29	16	2	7	11	4	19	3	7	26
-20	21	22	30	2	5	11	13	3	26	6	2	20	7	8	25	14	9	21	13	17	32	27	20	14	19	23	19	4	20	29	12	12	18	1	6	13	14	26	12	2	27	21	8	6	3	32	25	13	28
-5	11	23	16	10	9	19	6	21	6	27	17	22	14	26	3	10	16	22	31	27	18	19	5	2	5	2	24	15	6	18	16	12	26	16	17	17	16	22	27	13	20	17	5	31	23	4	4	9	32
-17	1	19	25	8	20	28	27	4	22	22	3	1	22	18	21	8	29	18	25	2	23	3	6	32	20	1	19	24	21	23	3	23	20	28	26	18	27	29	14	9	22	26	31	7	14	15	13	10	31
-3	12	31	23	32	21	10	3	4	27	26	15	28	32	14	22	7	22	9	18	21	22	1	16	8	17	5	6	2	3	14	2	28	24	11	21	27	6	16	31	2	9	3	6	10	32	4	23	7	16
-25	32	9	31	25	14	27	28	22	31	12	8	5	18	7	24	25	19	27	19	14	21	8	1	31	25	30	24	8	25	27	27	24	4	24	1	24	11	18	10	13	27	19	19	22	18	24	2	14	10
-25	17	26	8	13	29	17	24	6	5	22	18	17	26	10	18	25	29	29	11	28	26	3	24	31	22	28	20	1	11	22	1	4	27	14	11	22	7	27	12	8	8	14	6	15	25	26	22	8	23
-20	24	16	18	24	26	8	28	25	17	7	15	18	11	3	4	32	9	15	17	23	12	24	19	11	16	13	18	27	16	4	23	20	10	4	21	2	4	1	21	27	22	18	9	19	28	17	32	14	24
-1	12	29	2	10	18	14	29	25	28	8	27	1	17	8	21	3	32	19	25	9	16	28	6	25	31	6	24	13	13	5	15	17	14	11	22	5	17	8	19	21	3	12	22	28	26	14	8	24	28
-2	15	9	30	17	13	6	5	10	18	8	11	14	27	22	22	8	7	20	1	5	3	18	30	21	9	12	11	10	11	6	32	20	3	11	16	29	7	8	12	16	1	32	17	19	7	3	9	7	7
[…]
