+++ mcodes/results/datafiles/statematrices/s1500mcs1.txt
@@ -1,50 +1,0 @@
-3	3	3	3	3	3	3	3	3	3	3	3	3	4	4	4	4	4	4	4	4	4	4	18	18	18	18	16	16	31	16	16	30	30	30	30	30	27	27	27	27	27	27	27	27	27	27	27	27	3
-3	3	3	3	3	3	3	3	3	3	3	3	3	4	4	4	4	4	4	4	4	4	4	4	16	16	16	16	16	16	16	16	16	30	30	30	30	27	27	27	27	27	27	27	27	27	27	27	27	3
-3	3	3	3	3	21	3	3	3	3	3	3	3	6	4	4	4	4	4	4	4	4	4	4	16	16	16	16	16	16	16	16	16	16	30	30	30	27	27	27	27	27	27	27	27	27	27	27	23	3
-3	3	3	3	3	3	3	3	3	3	3	3	3	3	11	11	11	11	4	4	4	4	4	17	16	16	16	16	16	16	16	16	16	16	26	26	26	27	27	27	27	27	27	27	27	27	30	23	23	23
-23	3	3	3	3	3	3	3	3	3	3	3	3	18	11	11	11	11	11	4	4	4	4	16	16	16	16	16	16	16	16	16	16	16	26	26	26	26	27	27	27	27	27	27	27	23	23	23	23	23
-23	3	3	3	3	3	3	3	3	3	3	3	3	7	11	11	11	11	11	11	11	11	16	16	16	16	16	16	16	16	16	16	16	16	26	26	26	26	26	27	27	27	27	27	23	23	23	23	23	23
-23	3	3	3	3	3	3	3	3	3	3	3	22	22	22	22	11	11	11	11	11	11	16	16	16	16	16	16	16	16	16	16	16	16	26	26	26	26	2	2	2	2	2	2	23	23	23	23	23	23
-23	23	3	3	3	3	3	3	3	3	3	22	22	22	22	22	22	11	11	11	11	17	17	17	16	16	16	4	4	4	4	4	4	4	4	26	26	2	2	2	2	2	2	2	2	23	23	23	23	23
-23	23	3	3	3	3	3	3	3	3	22	22	22	22	22	22	22	22	11	11	17	17	17	17	17	16	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	23	23	23	23	23
-23	23	23	3	3	3	3	3	3	22	22	22	22	22	22	22	22	22	22	17	17	17	17	17	17	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	23	23	23	23
-23	23	23	3	3	3	3	3	22	22	22	22	22	22	22	22	22	22	22	17	17	17	17	17	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	23	23	23
-23	23	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	17	17	17	17	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	23	23	23
-23	9	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	22	17	17	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	7	7	23
-7	9	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	22	22	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	7	7	7	7
[…]
