+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using11_VERTICAL_intercepts.txt
@@ -1,4 +1,0 @@
-4.7122
-6.4786
-7.2404
-7.9267
