+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_1_incr =8.0069.txt
@@ -1,1 +1,0 @@
-8.0069
