+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using24_VERTICAL_intercepts.txt
@@ -1,4 +1,0 @@
-5.0556
-6.684
-7.7563
-8.4257

[…]
