### 2021-06-09 698f275
Updates to paper based on reviewer comments
1. Statement of need added
2. State of the field added
3. To account for the change, the paper has been slightly modified in wording and structure
4. Four additional references  have been added (1 for mathematical use: this was missing in the first submission), (3 for state of the field)
5. A back-up of the original paper has been made in paper_bu.md
