+++ paper.bib
@@ -1,8 +1,42 @@
+@article{JKMason2015,
+author = {Mason, J.K. and Lind, J. and Li, S.F. and Reed, B.W. and Kumar, M.},
+doi = {https://doi.org/10.1016/j.actamat.2014.08.063},
+issn = {1359-6454},
+journal = {Acta Materialia},
+pages = {155--166},
+title = {{Kinetics and anisotropy of the Monte Carlo model of grain growth}},
+volume = {82},
+year = {2015}
+}
+
+@book{JMarkKlien1999,
+author = {Mark Klien, J.},
+editor = {Sack, J.-R. and Urrita, J.},
+title = {{Chapter 11: Polygon decomposition, Handbook of Computational Geometry}},
+pages = {491--518},
+publisher = {Elsevier Science B. V},
+year = {1999}
+}
+
+@article{Quey2014,
+author = {Quey, R.},
+title = {{Neper: numerical descriptors of polycrystals (version 2.0), Available at http://neper.sourceforge.net/}},
+}
+
 @book{Baxter1989,
-author = {Baxter, Rodney J.},
+author = {Baxter, Rodney, J.},
 isbn = {0120831805},
 title = {{Exactly Solved Model in Statistica}},
 year = {1989}
+}
+
+@article{MAGroeber2014,
+author = {Groeber, M.A. and Jackson, M.A.},
+doi = {10.1186/2193-9772-3-5},
+journal = {Integrating Materials and Manufacturing Innovation},
+title = {{DREAM.3D: A Digital Representation Environment for the Analysis of Microstructure in 3D}},
+volume = {3:5},
+year = {2014}
 }
 
 @article{Weaire1996,
