+++ paper.bib
@@ -1,6 +1,6 @@
 @article{JKMason2015,
 author = {Mason, J.K. and Lind, J. and Li, S.F. and Reed, B.W. and Kumar, M.},
-doi = {https://doi.org/10.1016/j.actamat.2014.08.063},
+doi = {10.1016/j.actamat.2014.08.063},
 issn = {1359-6454},
 journal = {Acta Materialia},
 pages = {155--166},
