### 2021-06-10 #3190 @SunilAnandatheertha
@yangbai90 @littlelazy6 @ksyang2013 Hi I have incorporated changes accroding to your comments and replied here: [https://github.com/openjournals/joss-reviews/issues/3190#issuecomment-857788910]. Thank you.
