### 2021-06-22 #3190 @ksyang2013
Hi Kevin, thanks for your message. I am not able to finish the checklist and I keep getting the message "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." I guess I need to remove my assignment first.

In spite of the unfinished checklist, I am glad to share some general comments.
i) The generation and analysis of complex two-dimensional grain structures is indeed one important subject in computational materials science. The generated grain structures and corresponding analysis from the developed code are oriented at a macroscale modeling. It is fundamentally different from the existing codes or algorithms to build and/or model grain boundaries of materials at an atomic scale. Therefore, the application scene of the code should be clearly indicated and emphasized. This can be achieved by appropriately modifying the statements of need and summary in the article AND by adding application examples.
ii) The author is strongly encouraged to add some typical application examples that are best able to reflect the functionalities of the developed code, along with detailed instructions. A best practice could be that one application example (in one folder) only contains one major functionality.

Best,

--------------------------------------
Kesong YANG, Ph.D.

Associate Professor,
Department of NanoEngineering,
Program of Materials Science and Engineering,
Center for Memory and Recording Research (CMRR),
University of California San Diego
9500 Gilman Drive, Mail Code 0448
La Jolla, CA 92093-0448, USA

Office: SME 244H
Tel: (858) 534-2514
E-mail: ***@***.***
http://materials.ucsd.edu <http://materials.ucsd.edu/>

> On Jun 17, 2021, at 1:11 PM, Kevin Mattheus Moerman ***@***.***> wrote:
>
>
> @ksyang2013 <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_ksyang2013&d=DwMCaQ&c=-35OiAkTchMrZOngvJPOeA&r=45Q83Sy6H1-CqyoctWnD6w&m=0qzDKDROZFmUyYfSJ31t42lplWJC-l0pAU-Jz-oHAsU&s=_QQTWBc22Cn2UFxZ1jm0dDT0yQQfIYPd_pQM3LoRVzY&e=> can you please update me on review progress? Are you able to tick the boxes at the top of this issue? Thanks.
>
