### 2021-06-23 #3190 @ksyang2013
I still have trouble ticking these boxes. Can you remove my assignment and re-assign it?

Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account.

--------------------------------------
Kesong YANG, Ph.D.

Associate Professor,
Department of NanoEngineering,
Program of Materials Science and Engineering,
Center for Memory and Recording Research (CMRR),
University of California San Diego
9500 Gilman Drive, Mail Code 0448
La Jolla, CA 92093-0448, USA

Office: SME 244H
Tel: (858) 534-2514
E-mail: ***@***.***
http://materials.ucsd.edu <http://materials.ucsd.edu/>

> On Jun 23, 2021, at 11:04 AM, Kevin Mattheus Moerman ***@***.***> wrote:
>
>
> @ksyang2013 <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_ksyang2013&d=DwMFaQ&c=-35OiAkTchMrZOngvJPOeA&r=45Q83Sy6H1-CqyoctWnD6w&m=LcS5ELFRzeDZqVy9stlvCg0DIBX8ZtdVnhkb0Nl3Lb0&s=NMJ8vNJyVS4GCFdiXud2BfxwpLe5vTQCw2bRuOxUpzE&e=> thanks for the update, and sorry you had trouble ticking those boxes. Yes, I think it did time out. Can you accept this new invitation ☝️ ?
>
