### 2021-06-24 #3190 @littlelazy6
Thanks for the prompt revision. Just one issue with the documentation finishing level:
Could you please remind me where in wiki or document do you introduce how to define the initial VT generation parameters, such as:
1. The size and length_scale of the whole model (2D square)
2. The total number of grains, or mean and standard deviation used to define the grain size distribution
The above two is important when modelling the grain growth problem.
If applicable, please also define
3. How to define the initial VT aspect ratio or texture?
