+++ mcodes/VoronoiLimit.m
@@ -1,0 +1,993 @@
+function [V,C,XY]=VoronoiLimit(varargin)
+% --------------------------------------------------------------
+% [V,C,XY]=VoronoiLimit(x,y,additional_variables)
+% Provides the Voronoi decomposition of a set of (x,y) data, but with all
+% vertices limited to the boundary created by the data itself.
+% V contains the vertices of all voronoi-cells and C contains all vertices for each individual
+% voronoi-cell. That is: V(C{ij},:) will give you the vertices of the ij'th cell, corresponding 
+% to the ij'th data point, in two columns. The order of polygon vertices are given in a counter-clockwise
+% manner. XY contains updated xy coordinates as limited by any input boundaries.
+%
+% Addition variables:
+% 'bs_ext':  Describe an arbitrary external boundary by giving an xy matrix of size (n,1) where n are number of vertices.
+% 'bs_int':  Describe any number of arbitrary internal boundaries by giving a cell structure of M xy matrices of size
+%            (Ni,1) where M are number of internal boundaries and Ni are number of vertices in the respective boundaries. 
+%            When defining a single boundary the algorithm automatically converts the given matrix into a cell structure. 
+%            (See examples below).
+% 'figure':  output figure ('on'/'off'. Default='on').
+%
+% EXAMPLES
+% Example 0: Run with no input to see graphical example.
+%
+% Example 1: External and one internal boundary
+%            bs_int=[.2 .8 .8 .2;.6 .6 .2 .2]';
+%            bs_ext=[-.8 .5 1.80 -.8;-.05 1.7 -.05 -.05]';
+%            [X,Y] = meshgrid(-.5:.1:1.5, 0.1:.1:1.4);
+%            X=X(:);Y=Y(:);
+%            [V,C,XY]=VoronoiLimit(X,Y,'bs_ext',bs_ext,'bs_int',bs_int);
+%            [V,C,XY]=VoronoiLimit(X,Y,'bs_ext',bs_ext);
+%
+% Example 2: No external boundary and two internal boundaries
+%            bs_int=cell(2,1);
+%            bs_int{1}=[.2 .8 .8 .2;.6 .6 .2 .2]';
+%            bs_int{2}=[.2 .5 .7 .2;1 1 .7 .7]';
+%            [X,Y] = meshgrid(-.5:.1:1.5, 0.1:.1:1.4);
+%            X=X(:);Y=Y(:);
+%            [V,C,XY]=VoronoiLimit(X,Y,'bs_int',bs_int);
+% 
+% Example 3: As above but without figure output
+%            [V,C,XY]=VoronoiLimit(X,Y,'bs_int',bs_int,'figure','off');
+%
+% Requires the Polybool function of the mapping toolbox to run!.
+% I recommend the tool 'export_fig' for exporting figures. It can be found here:
+% http://www.mathworks.com/matlabcentral/fileexchange/23629-export-fig
+%
+% Made by: Jakob Sievers (PhD, Arctic Geophysics)
+% Contact: Jakob.Sievers@gmail.com
+% --------------------------------------------------------------
+warning('off','map:polygon:noExternalContours');
+version=[3 0 2 2];
+%% SETUP
+% USERTYPE
+[~,hostname]=system('hostname');
+user=0;
+if strcmp(hostname(1:end-1),'DESKTOP-PC4MSAH') || strcmp(hostname(1:end-1),'Sievers')
+    user=1;
+end
+% DETERMINE IF A MORE RECENT VERSION OF OOT IS AVAILABLE ON THE MATHWORKS FILEEXCHANGE
+try
[…]
