+++ README.md
@@ -1,26 +1,33 @@
 [![NumCodeLines](https://img.shields.io/tokei/lines/github/SunilAnandatheertha/PXO)]() 
-[![Last commit](https://img.shields.io/github/last-commit/SunilAnandatheertha/PXO)]() [![Total downloads](https://img.shields.io/github/downloads/SunilAnandatheertha/PXO/total)]() [![Size](https://img.shields.io/github/repo-size/SunilAnandatheertha/PXO)]()
+[![Last commit](https://img.shields.io/github/last-commit/SunilAnandatheertha/PXO)]()
+[![Total downloads](https://img.shields.io/github/downloads/SunilAnandatheertha/PXO/total)]()
+[![Size](https://img.shields.io/github/repo-size/SunilAnandatheertha/PXO)]()
+
+***
 
 # About PXO
-PXO: Poly-Crystal Operations, a free MATLAB based code-base to generate, analyze and export complex poly-crystalline multi-class spatial partitions
+### Introduction
+**PXO** stands for Poly-XTAL Operations. It is an open-source, MATLAB codebase for generating, analyzing and exporting complex space partitions such as multi-phase metallic grain structures. Poly-XTAL operations is named after the various operations which lead to its functionalities relating to poly-crystalline structures. It is useful for mathematicians, statistical mechanists, material scientists and computational geologiests. This wiki has pages for users, developers and collaborators.
 
-## Application fields
+Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [TESTCASES](https://github.com/SunilAnandatheertha/PXO/wiki/Test-cases) which you can try out while exploring or trying to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. Detailed information on exactly how to do this can be found in this wiki from the PXO-MTEX link provided in the sidebar of this page. You may use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to put forth your ideas, feature requests, documentation requests and bug reports. Development and all documentations by: [Sunil Anandatheertha](https://sunilanandatheertha.github.io/)
+
+### Application fields
 Materials science, geology, Statistical Mechanics, Space partitioning
 
+***
+
 # PXO Community
-### Sponsors list
-Sponsorer information will be listed here in the reverse chronological order
+### Usage
+Please start with [wiki page](https://github.com/SunilAnandatheertha/PXO/wiki). It will direct you. No installation is needed yet, to use PXO. Just add the folder to MATLAB path and start using.
 ### Discussions forum
 A [discussion](https://github.com/SunilAnandatheertha/PXO/discussions) forum has been set up, where the community members can share ideas, ask questions and interact with the developers and other users
-
-# Usage
-Please start with [wiki page](https://github.com/SunilAnandatheertha/PXO/wiki). It will direct you.
-### Installation
-No installation is needed yet, to use PXO. Just add the folder to MATLAB path and start using.
-## Contributing
+### Sponsors list
+Sponsorer information will be listed here in the reverse chronological order
+### Contributing
 Please go through contributing.md
-## Contributor credits
+### Contributor credits
 Contribute to get highlighted here
+
 
 ## License
 Refer licensing file for more information.
