+++ README.md
@@ -11,7 +11,7 @@
 
 Cite PXO as: 
 
-Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [TESTCASES](https://github.com/SunilAnandatheertha/PXO/wiki/Test-cases) which you can try out while exploring or trying to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. Detailed information on exactly how to do this can be found in this wiki from the PXO-MTEX link provided in the sidebar of this page. You may use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to put forth your ideas, feature requests, documentation requests and bug reports. Development and all documentations by: [Sunil Anandatheertha](https://sunilanandatheertha.github.io/)
+Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [Tutorials and Test Cases](https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases) which you can try out while learning to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. Detailed information on exactly how to do this can be found in this wiki from the PXO-MTEX link provided in the sidebar of this page. You may use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to put forth your ideas, feature requests, documentation requests and bug reports. Development and all documentations by: [Sunil Anandatheertha](https://sunilanandatheertha.github.io/)
 
 ### Application fields
 Materials science, geology, Statistical Mechanics, Space partitioning
