### 2021-06-30 #3190 @SunilAnandatheertha
@littlelazy6

In PXO, simulation of grain growth is carried out by using importance sampling Monte-Carlo (MC) schemes. Equivalent VT grain structures may also be produced. But, the equivalence is defined by the grain structure statistics as contained in the grain structure, as the grain structure evolves in the state-time space. The function of the codes for this, is to allow VT computation. Two types of VT computations can be generated. These are:

1.	VT over a user defined lattice. Currently, only 2D is available. Three types of lattices are possible, and these are random, rectangular and triangular. Together, they allow Voronoi tessellated random, rectangular and hexagonal grains respectively. Number of grains, size of grains, and aspect ratio of grains can be generated. Hexagonal grains with zigzag or armchair chirality may be generated. Rectangular grains can be made square grains as well, as the latter is subset of former. Long edge of a rectangular grain can be made to align either with x or y. These changes can be introduced by giving different values to a set of defining parameters for each case. A gradient may also be introduced if desired in each of these cases. Values of these parameters to obtain the grain structures have been detailed at https://github.com/SunilAnandatheertha/PXO/wiki/Voronoi-tessellation.

2.	VT equivalent of a MC generated grain structure at any of the allowed temporal slices. This is being refined and integrated. It will be included in PXO distribution in the subsequent releases.

However, as of now, PXO lacks the capability of generating either a MC grain structure or a VT grain structure to match a given grain structure statistics. This is a functionality present in the software DREAM3D and is being worked on to be introduced in PXO as well.
