### 2021-06-30 9d316de
Updates to paper and codes
Paper has been updated as per reviewer suggestions and comments.

VT grain structure code updated.

Previous version joss papers has been renamed and created
