### 2021-06-30 #3190 @SunilAnandatheertha
@littlelazy6

Initial texture can be defined in multiple ways.
1.	From assigning the EBSD acquired mean grain orientations to each grains by random sampling.
2.	From generating orientation samples for a give texture represented in MTEX.
3.	From generating a texture model in MTEX and then extracting the orientations from it.
These are included in a different repository and the relavant paper will be submitted soon. It will contain exporting the VT grain structures to ABAQUS and meshing in ABAQUS using custom python scripts.
