+++ mcodes/VTGS_V1.m
@@ -1,99 +1,213 @@
-close all
-clear all
-% rng(1)
-type = 'hex';
-switch type
-    case 'rect'
+close all; clear all
+%% GET USER INPUT DATA
+% 11111111111111111111111111111111111111111111111111111111
+prompt1 = {01, '01: UNIFORM-RECT-01'    , '';...
+           02, '02: UNIFORM-RECT-HEX-01', '';...
+           03, '03: UNIFORM-RANDOM-01'  , '';...
+           04, '04: UNIFORM-HEX-01'     , '';...
+           05, '05: GRAD-RECT-01-LOG'   , '';...
+           06, '06: GRAD-RECT-02-LOG'   , '';...
+           07, '07: GRAD-RECT-03-EXP'   , '';...
+           08, '08: GRAD-RAND-01-EXP'   , '';...
+           };
+dlgtitle = 'Type 1 below the option to use it.';
+LatType  = prompt1{~cellfun(@isempty,inputdlg(prompt1(:,2)',dlgtitle, [1 60], prompt1(:,3)')), 2};
+% 11111111111111111111111111111111111111111111111111111111
+prompt2 = {01, '01: VISUALIZE VORNOI CELLS: AREA COLOURING'         , 'ColourGrainsByArea';...
+           02, '02: VISUALIZE VORNOI CELLS: ASPECT RATIO COLOURING' , 'ColourGrainsByAspectRatio';...
+          };
+dlgtitle = 'Set flag for visualizing voronoi cells by area colouring';
+PlotGrainAspects = inputdlg(prompt2(:,2)',dlgtitle, [1 60], prompt2(:,3)');
+% 11111111111111111111111111111111111111111111111111111111
+if strcmp(PlotGrainAspects{1}, 'ColourGrainsByArea') == 1
+    prompt3 = {01, '01: BrightCyan (LOW) to Bright red (HIGH)', '1';...
+               02, '02: Black (LOW) to WHite (HIGH)'          , '';...
+               03, 'Colouring option - 3'   , '';...
+               04, 'Colouring option - 4'   , '';...
+               05, 'Colouring option - 5'   , '';...
+               06, 'Colouring option - 6'   , '';...
+               07, 'Colouring option - 7'   , '';...
+               };
+    dlgtitle = 'Select colouring option for area wise plotting';
+    AreaWiseGrainColours = prompt3{~cellfun(@isempty,inputdlg(prompt3(:,2)',dlgtitle, [1 100], prompt3(:,3)')), 2};
+end
+if strcmp(PlotGrainAspects{2}, 'ColourGrainsByAspectRatio') == 1
+    prompt3 = {01, '01: BrightCyan (LOW) to Bright red (HIGH)', '1';...
+               02, '02: Black (LOW) to WHite (HIGH)'          , '';...
+               03, 'Colouring option - 3'   , '';...
+               04, 'Colouring option - 4'   , '';...
+               05, 'Colouring option - 5'   , '';...
+               06, 'Colouring option - 6'   , '';...
+               07, 'Colouring option - 7'   , '';...
+               };
+    dlgtitle = 'Select colouring option for aspect ratio wise plotting';
+    AspectRatioWiseGrainColours = prompt3{~cellfun(@isempty,inputdlg(prompt3(:,2)',dlgtitle, [1 100], prompt3(:,3)')), 2};
+end
+% 22222222222222222222222222222222222222222222222222222222
+%% SELECT LATTICE TYPE, SET PARAMETERS, LIMITS AND MAKE THE LATTICE
+switch LatType
+    case '01: UNIFORM-RECT-01'
         startx  = 2;
[…]
