+++ mcodes/voronoigrainstructure_new.m
@@ -1,0 +1,205 @@
+DataFolder = 'C:\Users\anandats\OneDrive - Coventry University\coventry-thesis\MATLAB codes\Grain_Structure_data_Repository\';
+%----------------------------------------------
+rng(1)
+type = 'rect';
+switch type
+    case 'rect'
+        startx  = 2;
+        starty  = 0;
+        lengthx = 20;
+        widthy  = 4;
+        incr_i = 2.0;
+        incr_j = 0.5;
+        
+        bs_ext = [startx  starty;
+                  lengthx starty;
+                  lengthx widthy;
+                  startx  widthy;
+                  startx  starty];
+
+        xorig = startx+incr_i : incr_i : startx+lengthx-incr_i;
+        yorig = starty+incr_j : incr_j : starty+widthy -incr_j;
+        [x, y] = meshgrid(xorig, yorig);
+    case 'random'
+        startx  = 0;
+        starty  = 0;
+        lengthx = 5;
+        widthy  = 5;
+        Nopoints_i = 10;
+        Nopoints_j = 10;
+        xscale = 1.0; % not equal to zero
+        yscale = 1.0; % not equal to zero
+        origshiftx = +0.0;
+        origshifty = +0.0;
+
+        bs_ext = [startx  starty;
+                  lengthx starty;
+                  lengthx widthy;
+                  startx  widthy;
+                  startx  starty];
+
+
+        xorig = rand(Nopoints_i, Nopoints_j) - origshiftx;
+        yorig = rand(Nopoints_i, Nopoints_j) - origshifty;
+        %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+        x = xscale*lengthx*xorig;
+        y = yscale*widthy *yorig;
+    case 'hex'
+        %[xorig, yorig, xstart, ystart, xend, yend] = make_hex_grain_structure_v1();
+        %[xorig, yorig, xstart, ystart, xend, yend] = make_hex_grain_structure_v2();
+        x_lim = 5;
+        y_lim = 10;
+        h_dist = 1;
+        v_dist_factor = 1;
+        [Xfull, Yfull, xstart, ystart, xend, yend] = make_hex_grain_structure_v2(x_lim, y_lim, h_dist, v_dist_factor);
+        x = Xfull;
+        y = Yfull;
+        
+        startx  = xstart;
+        starty  = ystart;
+        lengthx = xend;
+        widthy  = yend;
+
+        bs_ext = [startx  starty;
+                  lengthx starty;
+                  lengthx widthy;
+                  startx  widthy;
+                  startx  starty];
+end
+%----------------------------------------------
+
+%----------------------------------------------
+randomfilenumber = num2str(randi(1e10));
+
+unscaledfilename_x = [DataFolder randomfilenumber '_xoriginal_unscaled_2dvor.txt'];
+unscaledfilename_y = [DataFolder randomfilenumber '_yoriginal_unscaled_2dvor.txt'];
+
+dlmwrite(unscaledfilename_x, xorig, '-append', 'delimiter', ',', 'roffset', 0)
+dlmwrite(unscaledfilename_y, yorig, '-append', 'delimiter', ',', 'roffset', 0)
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
[…]
