+++ mcodes/results/datafiles/statematrices/s1000mcs1.txt
@@ -1,0 +1,50 @@
+3	3	3	3	3	3	3	3	3	3	3	3	4	4	4	4	4	4	4	4	4	4	13	13	13	13	31	31	31	31	31	31	30	30	30	30	30	30	30	27	27	27	27	27	27	27	27	27	27	3
+3	3	3	3	3	3	3	3	3	3	3	3	3	4	4	4	4	4	4	4	4	13	13	13	13	13	31	31	31	31	31	16	16	30	30	30	30	30	27	27	27	27	27	27	27	27	27	27	27	3
+3	3	3	3	3	21	3	3	3	3	3	3	3	6	4	4	4	4	4	4	4	3	13	13	13	13	31	31	31	31	16	16	16	16	30	26	26	26	26	27	27	27	27	27	27	27	27	27	27	3
+3	3	3	3	3	3	3	3	3	3	3	3	3	3	11	11	11	11	11	11	3	3	3	17	12	13	16	16	16	16	16	16	16	16	26	26	26	26	26	26	27	27	27	27	27	27	30	23	23	23
+23	3	3	3	3	3	3	3	3	3	3	3	3	18	11	11	11	11	11	11	11	3	3	3	3	12	16	16	16	16	16	16	16	16	26	26	26	26	26	26	27	27	27	27	27	23	23	23	23	23
+23	23	3	3	3	3	3	3	3	3	3	3	3	7	11	11	11	11	11	11	11	3	3	3	3	26	26	16	16	16	16	16	16	16	26	26	26	26	26	26	27	27	27	27	23	23	23	23	23	23
+23	23	3	3	3	3	3	3	3	3	3	3	22	22	22	11	11	11	11	11	11	3	3	3	26	26	26	26	16	16	16	16	16	16	26	26	26	26	26	26	2	2	2	2	23	23	23	23	23	23
+23	23	3	3	3	3	3	3	3	3	3	22	22	22	22	22	11	11	11	11	11	3	3	26	26	26	26	26	16	16	16	16	16	24	24	26	26	26	26	2	2	2	2	2	23	23	23	23	23	23
+23	23	23	3	3	3	3	3	3	3	22	22	22	22	22	22	22	11	11	11	17	17	17	26	26	26	26	26	4	4	4	4	4	24	24	24	2	2	2	2	2	2	2	2	2	23	23	23	23	23
+23	23	23	23	3	3	3	3	3	22	22	22	22	22	22	22	22	22	11	17	17	17	17	17	26	26	4	4	4	4	4	4	4	4	24	24	2	2	2	2	2	2	2	2	2	2	23	23	23	23
+23	23	23	9	9	3	3	3	22	22	22	22	22	22	22	22	22	22	22	17	17	17	17	17	17	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	23	23	23	23
+23	23	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	22	17	17	17	17	17	17	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	23	23	23
+23	9	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	17	17	17	17	17	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	7	7	23
+9	9	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	17	17	17	17	17	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	7	7	7	7
[…]
