+++ mcodes/results/datafiles/graindata/grainsizes/A_A_AGSvsMCS_M_intincr_1_2_and_3 =7.9615.txt
@@ -1,0 +1,1 @@
+7.9615
