+++ mcodes/results/datafiles/coordinates/y.txt
@@ -1,0 +1,50 @@
+-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24	-24
+-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23	-23
+-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22	-22
+-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21	-21
+-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20	-20
+-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19	-19
+-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18	-18
+-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17	-17
[…]

[…]
