+++ README.md
@@ -9,7 +9,7 @@
 ### Introduction
 **PXO** stands for Poly-XTAL Operations. It is an open-source, MATLAB codebase for generating, analyzing and exporting complex space partitions such as multi-phase metallic grain structures. Poly-XTAL operations is named after the various operations which lead to its functionalities relating to poly-crystalline structures. It is useful for mathematicians, statistical mechanists, material scientists and computational geologiests. This wiki has pages for users, developers and collaborators.
 ***
-Cite PXO as: 
+Cite PXO as: Anandatheertha, S., (2021). PXO (Poly-XTAL Operations): MATLAB codebase to generate and analyse complex 2D grain structures. Journal of Open Source Software, 6(62), 3190. https://doi.org/10.21105/joss.03190
 ***
 Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [Tutorials and Test Cases](https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases) which you can try out while learning to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. Detailed information on exactly how to do this can be found in this wiki from the PXO-MTEX link provided in the sidebar of this page. You may use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to put forth your ideas, feature requests, documentation requests and bug reports. Development and all documentations by: [Sunil Anandatheertha](https://sunilanandatheertha.github.io/)
 
@@ -19,17 +19,7 @@ Materials science, geology, Statistical Mechanics, Space partitioning
 ***
 
 # PXO Community
-### Usage
-Please start with [wiki page](https://github.com/SunilAnandatheertha/PXO/wiki). It will direct you. No installation is needed yet, to use PXO. Just add the folder to MATLAB path and start using.
-### Discussions forum
-A [discussion](https://github.com/SunilAnandatheertha/PXO/discussions) forum has been set up, where the community members can share ideas, ask questions and interact with the developers and other users
-### Sponsors list
-Sponsorer information will be listed here in the reverse chronological order
-### Contributing
-Please go through contributing.md
-### Contributor credits
-Contribute to get highlighted here
-
+Please start with [wiki page](https://github.com/SunilAnandatheertha/PXO/wiki). It will direct you. No installation is needed yet, to use PXO. Just add the folder to MATLAB path and start using. A [discussion](https://github.com/SunilAnandatheertha/PXO/discussions) forum has been set up, where the community members can share ideas, ask questions and interact with the developers and other users. Sponsorer information will be listed here in the reverse chronological order. Please go through contributing.md to see how you may contribute. Contributer credits will be listed in detail here.
 
 ## License
 Refer licensing file for more information.
