+++ README.md
@@ -1,25 +1,19 @@
-[![NumCodeLines](https://img.shields.io/tokei/lines/github/SunilAnandatheertha/PXO)]() 
+***
+### Introduction
+**PXO** stands for Poly-XTAL Operations. It is an open-source, MATLAB codebase for generating, analyzing and exporting complex space partitions such as multi-phase metallic grain structures. Poly-XTAL operations is named after the various operations which lead to its functionalities relating to poly-crystalline structures. It is useful for mathematicians, statistical mechanists, material scientists and computational geologiests. Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [Tutorials and Test Cases](https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases) which you can try out while learning to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. You may use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to as questions/clarifications, put forth your ideas, feature requests, documentation requests and bug reports. 
+
+Development and all documentations by: [Sunil Anandatheertha](https://sunilanandatheertha.github.io/). Please start with [wiki page](https://github.com/SunilAnandatheertha/PXO/wiki). It will direct you. To use PXO, no installation is needed yet. Just add the folder to MATLAB path and start using. A [discussion](https://github.com/SunilAnandatheertha/PXO/discussions) forum has been set up, where the community members can share ideas, ask questions and interact with the developers and other users. Sponsorer information will be listed here in the reverse chronological order. Please go through contributing.md to see how you may contribute. Contributer credits will be listed in detail here.
+
+***
+Please cite PXO as: Anandatheertha, S., (2021). PXO (Poly-XTAL Operations): MATLAB codebase to generate and analyse complex 2D grain structures. Journal of Open Source Software, 6(62), 3190. https://doi.org/10.21105/joss.03190
+
+***
+## License
+Refer licensing file for more information.
+
+***
+
 [![Last commit](https://img.shields.io/github/last-commit/SunilAnandatheertha/PXO)]()
 [![Total downloads](https://img.shields.io/github/downloads/SunilAnandatheertha/PXO/total)]()
 [![Size](https://img.shields.io/github/repo-size/SunilAnandatheertha/PXO)]()
-
-***
-
-# About PXO
-### Introduction
-**PXO** stands for Poly-XTAL Operations. It is an open-source, MATLAB codebase for generating, analyzing and exporting complex space partitions such as multi-phase metallic grain structures. Poly-XTAL operations is named after the various operations which lead to its functionalities relating to poly-crystalline structures. It is useful for mathematicians, statistical mechanists, material scientists and computational geologiests. This wiki has pages for users, developers and collaborators.
-***
-Cite PXO as: Anandatheertha, S., (2021). PXO (Poly-XTAL Operations): MATLAB codebase to generate and analyse complex 2D grain structures. Journal of Open Source Software, 6(62), 3190. https://doi.org/10.21105/joss.03190
-***
[…]
