### 2021-07-06 #3190 @littlelazy6
> @littlelazy6
>
> Initial texture can be defined in multiple ways.
>
> 1. From assigning the EBSD acquired mean grain orientations to each grains by random sampling.
> 2. From generating orientation samples for a give texture represented in MTEX.
> 3. From generating a texture model in MTEX (for Voronoi tessellations) and then extracting the orientations from it.
>    These are included in a different repository and the relavant paper will be submitted soon. It will contain exporting the VT grain structures to ABAQUS and meshing in ABAQUS using custom python scripts.

Thanks for the detailed response. I have checked and convinced the capability. Look forwards to the filter (plugin) working with DREAM3D for more advanced application such as 3D models.
@Kevin-Mattheus-Moerman I have updated the review checkpoints. Please let me know if there are any other tasks on my side.
