### 2021-07-07 #3190 @ksyang2013
Thank you! It works now!

--------------------------------------
Kesong YANG, Ph.D.

Associate Professor,
Department of NanoEngineering,
Program of Materials Science and Engineering,
Center for Memory and Recording Research (CMRR),
University of California San Diego
9500 Gilman Drive, Mail Code 0448
La Jolla, CA 92093-0448, USA

Office: SME 244H
Tel: (858) 534-2514
E-mail: ***@***.***
http://materials.ucsd.edu <http://materials.ucsd.edu/>

> On Jul 7, 2021, at 6:25 AM, Kevin Mattheus Moerman ***@***.***> wrote:
>
>
> @ksyang2013 <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_ksyang2013&d=DwMCaQ&c=-35OiAkTchMrZOngvJPOeA&r=45Q83Sy6H1-CqyoctWnD6w&m=wi-ooKpFOoTAM7hN4kalBQUlvooT2k_pCDwL5j-cS4I&s=HIs683BIuZSrfydQT85tMtDgyrwYyC4M_UHqwBgjFa0&e=> I have sent you another inviting hoping this might work. If it doesn't work we will look into this. As an alternative you can also copy the components at the top of this issue and tick them in a normal comment window here.
>
