### 2021-07-07 #3190 @ksyang2013
@SunilAnandatheertha <https://github.com/SunilAnandatheertha>

Thanks for updates. I am checking the following items:

 Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Even it is stated in README.md <https://github.com/SunilAnandatheertha/PXO#readme>, it is still desired to put an instruction like “To use PXO, no installation is needed yet. Just add the folder to MATLAB path and start using.” In a document like “INSTALL.txt”. Make it easy to follow ...

 Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
Have you provided a script that can help check the authors’ test?  Maybe one simple test with core function is enough.

I understood you have put "Test Case 01 <https://github.com/SunilAnandatheertha/PXO/wiki/PXO-to-MTEX-to-ABAQUS-01>:” in wiki, but I did meet some problems when running, like below:

50x50.||.m:2000/2000.
Elapsed time is 0.592793 seconds.
currentplot: 500
Error using matlab.graphics.internal.name (line 112)
Cannot create output file
'/Users/xxx/Documents/MATLAB/mcodes/results/plots/microstructure_withgb/500mcs.jpeg'.

Error in print (line 71)
pj = matlab.graphics.internal.name( pj );

Error in plotgrainstructure2d (line 80)
            print('-djpeg100',strcat(pwd, '\results\plots\microstructure_withgb','\',
            strcat(num2str(rset*txtwriteint),'mcs'),'.jpeg'))

Error in gss2d (line 15)
[CFN] = plotgrainstructure2d(1, 1, 1, 998);

Thanks,

--------------------------------------
Kesong YANG, Ph.D.

Associate Professor,
Department of NanoEngineering,
Program of Materials Science and Engineering,
Center for Memory and Recording Research (CMRR),
University of California San Diego
9500 Gilman Drive, Mail Code 0448
La Jolla, CA 92093-0448, USA

Office: SME 244H
Tel: (858) 534-2514
E-mail: ***@***.***
http://materials.ucsd.edu <http://materials.ucsd.edu/>

> On Jun 30, 2021, at 6:13 AM, Sunil Anandatheertha ***@***.***> wrote:
>
>
> @ksyang2013 <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_ksyang2013&d=DwMCaQ&c=-35OiAkTchMrZOngvJPOeA&r=45Q83Sy6H1-CqyoctWnD6w&m=WMyrn0FMMABF_U2jDkdLhSC8uguLN6nAnCd28lLGBcM&s=kArBlyX2-x9xdszlKqNJh_JWnNfP5pTlyj1Ywm8RD7k&e=> Paper has been updated accordingly. Tutorials has been added and the test case is retained (here: https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_SunilAnandatheertha_PXO_wiki_Tutorials-2Dand-2Dtest-2Dcases&d=DwMCaQ&c=-35OiAkTchMrZOngvJPOeA&r=45Q83Sy6H1-CqyoctWnD6w&m=WMyrn0FMMABF_U2jDkdLhSC8uguLN6nAnCd28lLGBcM&s=-Wo3XvEytdOlWNL247BIwFol5UlFWvGlW5j_ejcrR3o&e=>)
>
