### 2021-07-08 6fc881f
Installation requirements
Update as suggested by JOSS paper reviewer @ksyang on GitHub
