+++ mcodes/Build__CTF__ARTIFICIAL.m
@@ -1,0 +1,181 @@
+function Build__CTF__ARTIFICIAL(phi1, psi, phi2, PHASEMATRIX, CTF_FILENAME)
+% CTF_FILENAME
+%------------------------------------------------------------------------------------------------------------
+global Lattice EBSD_LIKE_DATA
+%------------------------------------------------------------------------------------------------------------
+x                = Lattice.size.x;
+y                = Lattice.size.y;
+x                = x - min(min(x));
+y                = y - min(min(y));
+xincr            = Lattice.size.i_incr;
+yincr            = Lattice.size.j_incr;
+LatticeSize      = size(x);
+NumLatticeSites  = numel(x);
+%------------------------------------------------------------------------------------------------------------
+Ctf_Build.line01 = 'Channel Text File';
+Ctf_Build.line02 = 'Prj	Project 1';
+Ctf_Build.line03 = 'Author	';
+Ctf_Build.line04 = 'JobMode	Grid';
+Ctf_Build.line05 = ['XCells	' num2str(LatticeSize(1))];
+Ctf_Build.line06 = ['YCells	' num2str(LatticeSize(2))];
+Ctf_Build.line07 = ['XStep	' num2str(xincr)];
+Ctf_Build.line08 = ['YStep	' num2str(yincr)];
+Ctf_Build.line09 = 'AcqE1	90.0000';
+Ctf_Build.line10 = 'AcqE2	90.0000';
+Ctf_Build.line11 = 'AcqE3	-90.0000';
+
+hasvoids = 0;
+
+Ctf_Build.line12 = 'Euler angles refer to Sample Coordinate system (CS0)!	Mag	100.0000	Coverage	74	Device	0	KV	20.0000	TiltAngle	70.0000	TiltAxis	0	DetectorOrientationE1	358.7268	DetectorOrientationE2	92.6504	DetectorOrientationE3	358.2069	WorkingDistance	10.6012	InsertionDistance	202.8906';
+if hasvoids == 1
+    Ctf_Build.line13a = 'Phases	3';
+else
+    Ctf_Build.line13a = 'Phases	2';
+end
+Ctf_Build.line14a = '4.050;4.050;4.050	90.000;90.000;90.000	Aluminium	11	225			Cryogenics18,54-55';
+Ctf_Build.line14b = '2.000;2.000;2.000	90.000;90.000;90.000	PARTICLES	11	225			Cryogenics18,54-55';
+if hasvoids == 1
+    Ctf_Build.line14c = '1.000;1.000;1.000	90.000;90.000;90.000	HOLEVOIDS	11	225			Cryogenics18,54-55';
+end
+Ctf_Build.line15 = 'Phase	X	Y	Bands	Error	Euler1	Euler2	Euler3	MAD	BC	BS';
+%------------------------------------------------------------------------------------------------------------
+%------------------------------------------------------------------------------------------------------------
+Phase1 = reshape(PHASEMATRIX, NumLatticeSites, 1)
+x1     = reshape(x', NumLatticeSites, 1);
+y1     = reshape(y', NumLatticeSites, 1);
+Bands1 = reshape(05 + randi(5, LatticeSize), NumLatticeSites, 1);
+Error1 = reshape(00 + randi(3, LatticeSize), NumLatticeSites, 1);
+phi1_1 = reshape(phi1, NumLatticeSites, 1);
+psi_1  = reshape(psi , NumLatticeSites, 1);
+phi2_1 = reshape(phi2, NumLatticeSites, 1);
+MAD_1  = reshape(rand(LatticeSize), NumLatticeSites, 1);
[…]
