### 2021-07-08 bf45c37
Updated Build___INPUT_PARAMS___Struct
