### 2021-07-08 64e1ed4
Updated PXO plotting of grain structure - fast
