### 2021-07-08 65b385e
Updated initialize mc date codes
