### 2021-07-08 95be777
New solver alg
KERNEL__LOOP_MC_2D____ALGORITHM_0302
