+++ mcodes/KERNEL__LOOP_MC_2D____ALGORITHM_0302.m
@@ -1,0 +1,114 @@
+function KERNEL__LOOP_MC_2D____ALGORITHM_0302(initialmcs, finalmcs, consider_energy, E, delE, ham, SZS1P1, SZS1P2, SZS1P3, vf, skipmatrix, ConsoleDisplay)
+%-------------------------------------------------------------
+global MC_Param Lattice MC_Loop CMDL_display File_Fold_Operations
+%-------------------------------------------------------------
+sz1             = Lattice.size.sz1;
+sz2             = Lattice.size.sz2;
+s               = Lattice.orientations.s;
+q               = Lattice.q;
+ea              = Lattice.s__MATLAB_Indices__BounCond_Wrapped;
+WM              = MC_Param.WeightMatrix;
+Consider_Energy = MC_Param.Consider_Energy;
+txtwriteint     = MC_Loop.DataOperation.txtwriteint;
+nof             = File_Fold_Operations.writedlm.s.nof;
+%-------------------------------------------------------------
+start = initialmcs;
+%-------------------------------------------------------------
+%0000000000000000000000000000000000000000000000000000000000000000000000000000
+% DEFGINE THE FUNCTION WHICH DEFINES THE SPATIAL DISTRIBUTION OF
+% "J_by_k_by_T"
+% sz1
+% sz2
+[JKT_FIELD] = ConstructJKT_field_v1([0 5 0 5], 1:sz1, 1:sz2, 'x-min_to_x+max', {'power', 1, -0.1, 'norm_max_yes'});
+
+
+% J_by_k_by_T = 0.0;
+%0000000000000000000000000000000000000000000000000000000000000000000000000000
+%-------------------------------------------------------------
+Transition_Probability = exp(-JKT_FIELD);
+
+figure
+contourf(Transition_Probability); colorbar
+%-------------------------------------------------------------
+for ms = start:finalmcs
+    NS_full = floor(1+q*rand(sz1,sz2));
+    for pt = 1:numel(s)
+        if skipmatrix(pt) == 0
+            pta  = 2*ceil(pt/sz1) + SZS1P1 + pt;
+            esm1 = s(ea(pta-SZS1P3));
+            esm2 = s(ea(pta-SZS1P2));
+            esm3 = s(ea(pta-SZS1P1));
+            esm4 = s(ea(pta-1));
+            esm5 = s(ea(pta));
+            esm6 = s(ea(pta+1));
+            esm7 = s(ea(pta+SZS1P1));
+            esm8 = s(ea(pta+SZS1P2));
+            esm9 = s(ea(pta+SZS1P3));
+%             energy1 = 8 - ((esm5==esm1) + (esm5==esm4) + (esm5==esm7) + (esm5==esm2) + (esm5==esm8) + (esm5==esm3) + (esm5==esm6) + (esm5==esm9));
+%             energy1 = esm5==esm1 + esm5==esm4 + esm5==esm7 + esm5==esm2 + esm5==esm8 + esm5==esm3 + esm5==esm6 + esm5==esm9;
+            NS = NS_full(pt);
+                      
+%             states = [esm1; esm2; esm3; esm4; esm6; esm7; esm8; esm9];
+%             NS = states(floor(1+8*rand(1)));
+%             states = [esm1; esm2; esm3; esm4; esm6; esm7; esm8; esm9];
+%             NS = states(floor(1+8*rand(1)));
[…]
