### 2021-07-08 e9a3d94
create temperature field
