### 2021-07-08 3e4cabe
updated hamiltonian minimization function
