### 2021-07-08 ff77599
Updated alg 1 and created new algs
