### 2021-07-08 fa0697a
Calculate grain structure data using mtex
