### 2021-07-08 da9ceee
mcode for setting plot properties
