+++ mcodes/SetPlottingProperties01.m
@@ -1,0 +1,120 @@
+function SetPlottingProperties01(WHICH_CASE, AX_LIM, FONTSIZES, DATASET1)
+% FONTSIZES  = [12 14 12]; % GCA font size, label font size, legend font size
+switch WHICH_CASE
+    case 'jsfhj547'
+        xaxmin = AX_LIM(1); xaxmax = AX_LIM(2); 
+        yaxmin = AX_LIM(3); yaxmax = AX_LIM(4);
+        hxlab  = xlabel('MC sim. time');
+        hylab  = ylabel('Peak position, unit^2');
+        hleg   = legend('Peak position', 'location', 'southwest');
+        set(gca, 'fontsize', FONTSIZES(1))
+        set(hleg, 'box', 'off')
+        set([hxlab hylab], 'fontsize', FONTSIZES(2))
+        set(hleg, 'fontsize', FONTSIZES(3))
+        axis([xaxmin xaxmax yaxmin yaxmax])
+        grid on
+        ax = gca;
+        MCSlices = DATASET1;
+        ax.XTick = 10.^[0:log10(MCSlices(end))];
+        ax.YTick = linspace(yaxmin, yaxmax, 6);
+        tiy = get(gca,'ytick')'; set(gca,'yticklabel',num2str(tiy,'%.1f'))
+    case 'skhvbdjef683'
+        xaxmin = AX_LIM(1); xaxmax = AX_LIM(2);
+        yaxmin = AX_LIM(3); yaxmax = AX_LIM(4);
+        hxlab  = xlabel('MC sim. time');
+        hylab  = ylabel('Peak height, count');
+        hleg   = legend('Peak height', 'location', 'northeast');
+        set(gca, 'fontsize', FONTSIZES(1))
+        set(hleg, 'box', 'off')
+        set([hxlab hylab], 'fontsize', FONTSIZES(2))
+        set(hleg, 'fontsize', FONTSIZES(3))
+        axis([xaxmin xaxmax yaxmin yaxmax])
+        grid on
+        ax = gca;
+        MCSlices = DATASET1;
+        ax.XTick = 10.^[0:log10(MCSlices(end))];
+        ax.YTick = linspace(yaxmin, yaxmax, 6);
+        tiy = get(gca,'ytick')'; set(gca,'yticklabel',num2str(tiy,'%.1f'))        
+    case 'fbkndlgf9846'
+        xaxmin = AX_LIM(1);
+        xaxmax = AX_LIM(2);
+        yaxmin = AX_LIM(3);
+        yaxmax = AX_LIM(4);
+        hxlab  = xlabel('MC sim. time');
+        hylab  = ylabel('Peak width, unit^2');
+        hleg   = legend('Peak width', 'location', 'northeast');
+        set(gca, 'fontsize', FONTSIZES(1))
+        set(hleg, 'box', 'off')
+        set([hxlab hylab], 'fontsize', FONTSIZES(2))
+        set(hleg, 'fontsize', FONTSIZES(3))
+        axis([xaxmin xaxmax yaxmin yaxmax])
+        grid on
+        ax = gca;
+        MCSlices = DATASET1;
+        ax.XTick = 10.^[0:log10(MCSlices(end))];
+        ax.YTick = linspace(yaxmin, yaxmax, 6);
+        tiy = get(gca, 'ytick')'; set(gca, 'yticklabel',num2str(tiy, '%.1f'))
+    case 'GrainAreaDistribution_hist_plotMarkup_sjkhvbke78'
+        xaxmin = AX_LIM(1);
+        xaxmax = AX_LIM(2);
+        yaxmin = AX_LIM(3);
[…]
