+++ mcodes/ExtractGS_Data_Statistics_V1.m
@@ -1,0 +1,117 @@
+function [GrArea_PeakPosition_TRIALS, GrArea_PeakHeight_TRIALS, GrArea_PeakWidth_TRIALS,...
+NumNeigh_PeakPosition_TRIALS, NumNeigh_PeakHeight_TRIALS, NumNeigh_PeakWidth_TRIALS,...
+EqRad_PeakPosition_TRIALS, EqRad_PeakHeight_TRIALS, EqRad_PeakWidth_TRIALS,...
+EqPer_PeakPosition_TRIALS, EqPer_PeakHeight_TRIALS, EqPer_PeakWidth_TRIALS,...
+ShFac_PeakPosition_TRIALS, ShFac_PeakHeight_TRIALS, ShFac_PeakWidth_TRIALS,...
+AsRat_PeakPosition_TRIALS, AsRat_PeakHeight_TRIALS, AsRat_PeakWidth_TRIALS] = ExtractGS_Data_Statistics_V1(TOTAL_TRIALS, finalMC_TimeSlice,...
+                          GrainSizes_mtex_TRIALS,...
+                          numNeighbors_mtex_TRIALS,...
+                          equivalentRadius_mtex_TRIALS,...
+                          equivalentPerimeter_mtex_TRIALS,...
+                          ShapeFactor_mtex_TRIALS,...
+                          AspectRatio_mtex_TRIALS)
+
+close all
+disp('-----------------------------------------')
+GrArea_PeakPosition_TRIALS   = cell(1, TOTAL_TRIALS);GrArea_PeakHeight_TRIALS     = cell(1, TOTAL_TRIALS);GrArea_PeakWidth_TRIALS      = cell(1, TOTAL_TRIALS);
+NumNeigh_PeakPosition_TRIALS = cell(1, TOTAL_TRIALS);NumNeigh_PeakHeight_TRIALS   = cell(1, TOTAL_TRIALS);NumNeigh_PeakWidth_TRIALS    = cell(1, TOTAL_TRIALS);
+EqRad_PeakPosition_TRIALS    = cell(1, TOTAL_TRIALS);EqRad_PeakHeight_TRIALS      = cell(1, TOTAL_TRIALS);EqRad_PeakWidth_TRIALS       = cell(1, TOTAL_TRIALS);
+EqPer_PeakPosition_TRIALS    = cell(1, TOTAL_TRIALS);EqPer_PeakHeight_TRIALS      = cell(1, TOTAL_TRIALS);EqPer_PeakWidth_TRIALS       = cell(1, TOTAL_TRIALS);
+ShFac_PeakPosition_TRIALS    = cell(1, TOTAL_TRIALS);ShFac_PeakHeight_TRIALS      = cell(1, TOTAL_TRIALS);ShFac_PeakWidth_TRIALS       = cell(1, TOTAL_TRIALS);
+AsRat_PeakPosition_TRIALS    = cell(1, TOTAL_TRIALS);AsRat_PeakHeight_TRIALS      = cell(1, TOTAL_TRIALS);AsRat_PeakWidth_TRIALS       = cell(1, TOTAL_TRIALS);
+disp('-----------------------------------------')
+PeakFitType_GrainSizes_mtex = 25; % Log-Normal
+disp('-----------------------------------------')
+
+for trial_count = 1:TOTAL_TRIALS
+    disp('Initialising Gr.Strc. property extraction')
+    %------------------------------------
+    GrArea_PeakPosition   = zeros(finalMC_TimeSlice, 1);GrArea_PeakHeight   = zeros(finalMC_TimeSlice, 1);GrArea_PeakWidth   = zeros(finalMC_TimeSlice, 1);
[…]
