### 2021-07-08 19b6652
including num2clip 3rd party code
allows to copy data to clipboard to enable copying data to excel sheet
