+++ mcodes/Calculate_GS_Data_using_mtex.m
@@ -1,0 +1,111 @@
+function [GrainSizes_mtex, numNeighbors_mtex, equivalentRadius_mtex, equivalentPerimeter_mtex, ShapeFactor_mtex] = Calculate_GS_Data_using_mtex()
+
+global MC_Param MC_Loop
+%---------------------------------------------------------
+% rset = 4;
+%------------------------------------------------------
+m            = MC_Param.Num_MC_Steps;
+txtwriteint  = MC_Loop.DataOperation.txtwriteint;
+%---------------------------------------------------------
+TimeSteps    = txtwriteint:txtwriteint:m;
+N_Time_Slices = numel(TimeSteps);
+%---------------------------------------------------------
+CTF_FOLDER = [pwd filesep 'results\datafiles' filesep 'CTF_FILES_MCSTIME_SLICES'];
+%---------------------------------------------------------
+%---------------------------------------------------------
+if ~exist(CTF_FOLDER, 'dir')
+    mkdir(CTF_FOLDER)
+end
+%---------------------------------------------------------
+GrainSizes_mtex          = cell(N_Time_Slices,1);
+numNeighbors_mtex        = cell(N_Time_Slices,1);
+equivalentRadius_mtex    = cell(N_Time_Slices,1);
+equivalentPerimeter_mtex = cell(N_Time_Slices,1);
+ShapeFactor_mtex         = cell(N_Time_Slices,1);
+%---------------------------------------------------------
+for rset = 1:N_Time_Slices % Time slice level
+    disp('---------------------------------------------------------')
+    fprintf(['Working on MC time slice # ' num2str(TimeSteps(rset)) '\n'])
+    %---------------------------------------------------------
+    % LOAD THE STATE MATRIX IN THIS TIME SLICE
+    statematrix = dlmread([pwd filesep 'results\datafiles\statematrices' filesep 's' num2str(TimeSteps(rset)) 'mcs1' '.txt']);
+    %---------------------------------------------------------
+    % MAKE THE CTF FILE FOR THIS TIME SLICE
+    disp('Building CTF file')
+    [phi1, psi, phi2, PHASEMATRIX] = Associate_EA_all_sites_ARTIFICIAL('2d', statematrix);
+    CTF_FILENAME = [CTF_FOLDER filesep 'CTF_MCS_' num2str(TimeSteps(rset)) '.ctf'];
+    Build__CTF__ARTIFICIAL(phi1, psi, phi2, PHASEMATRIX, CTF_FILENAME)
+    %---------------------------------------------------------
+    %---------------------------------------------------------
+    %---------------------------------------------------------
+    %---------------------------------------------------------
+    % USE MTEX TO MAKE GRAIN STRUCTURE ANALYSIS
+    disp('Loading CTF file from mtex')
+    ebsd_Gr_STR = loadEBSD_ctf(CTF_FILENAME,'convertSpatial2EulerReferenceFrame');
+    ebsd_Gr_STR('PARTICLES').color = [1 0 0];
+    % + + + + + + + + + + + + + + + 
+    disp('Calculating grains and properties')
+    grains = calcGrains(ebsd_Gr_STR('Aluminium'),'angle',0.1*degree);
+    number_of_grains = size(grains,1);
+    % + + + + + + + + + + + + + + + 
+% %     disp('plotting grain structure')
+% %     figure; hold on
+% %     plot(grains,'translucent',1,'micronbar', 'off', 'coordinates','off'); hold on
+% %     plot(ebsd_Gr_STR('PARTICLES'),'coordinates','off'); hold on
+% %     legend off
+% %     %plot(grains.boundary,'color','k','linewidth',1.5)
[…]
