### 2021-07-08 639738d
Associate orientations to real ctf files
