### 2021-07-08 48177a8
build ctf files of real gs
