### 2021-07-08 6b287dd
Updated solver codes and intercept gs size calculation codes
