### 2021-07-08 c4dfbc0
results from sample run
