+++ mcodes/txtfiledet.txt
@@ -1,4 +1,1 @@
-500	2
-1000	2
-1500	2
-2000	2
+100	2
