+++ mcodes/results/colormatrix.txt
@@ -1,32 +1,10 @@
-0.83706	0.10779	0.57768
-0.24277	0.8689	0.34238
-0.50573	0.0032652	0.13547
-0.46839	0.51859	0.50839
-0.8049	0.55082	0.64626
-0.48547	0.11595	0.96752
-0.047994	0.20004	0.76195
-0.29891	0.86222	0.183
-0.72163	0.11697	0.23966
-0.54076	0.8247	0.95701
-0.71301	0.19309	0.758
-0.89043	0.99555	0.19292
-0.036948	0.47141	0.70195
-0.16157	0.54727	0.83347
-0.20618	0.082274	0.43625
-0.54819	0.051293	0.0094767
-0.6875	0.20742	0.66811
-0.34233	0.16079	0.069735
-0.8698	0.45106	0.2378
-0.8792	0.62995	0.097031
-0.19012	0.72633	0.87047
-0.53183	0.45605	0.83404
-0.98953	0.37693	0.23516
-0.057573	0.064415	0.75438
-0.083431	0.53834	0.50836
-0.41763	0.45104	0.42116
-0.85461	0.40236	0.21872
-0.39859	0.073837	0.22166
-0.0099489	0.19382	0.090802
-0.64554	0.60843	0.24882
-0.2938	0.91508	0.58965
-0.14993	0.63152	0.88221
+0.82911	0.84829	0.23073
+0.73298	0.15117	0.40413
+0.36253	0.58585	0.7406
+0.99035	0.80131	0.0034284
+0.76143	0.15534	0.95716
+0.10108	0.15286	0.61593
+0.23972	0.86705	0.18572
+0.27225	0.28858	0.51468
+0.055229	0.89345	0.72977
+0.046697	0.32331	0.44833
