+++ mcodes/simparameters/solset.txt
@@ -1,3 +1,3 @@
 0
-500
+100
 10
