+++ mcodes/simparameters/simpar.txt
@@ -1,2 +1,2 @@
-32
-2000
+10
+100
