+++ mcodes/simparameters/initialfinalmcs.txt
@@ -1,2 +1,2 @@
 1
-2000
+100
