+++ mcodes/results/datafiles/e/zener_allover.txt
@@ -1,25 +1,25 @@
-388
-655
-1584
-253
-1289
-1118
-127
-1915
-238
-846
-1448
-2128
-653
-2026
-1092
-2070
-2304
-1096
-656
-1154
-238
-1647
-1581
-1242
-1176
+244
+780
+817
+1024
+746
+611
+2382
+251
+796
+2042
+1917
+401
+89
+274
+762
+912
+1300
+2040
+2323
+1939
+96
+2102
+2486
+2061
+731
