### 2021-07-08 c3702c3
Added link to installation instructions
in the readme.md file.
