+++ README.md
@@ -1,6 +1,6 @@
 ***
 ### Introduction
-**PXO** stands for Poly-XTAL Operations. It is an open-source, MATLAB codebase for generating, analyzing and exporting complex space partitions such as multi-phase metallic grain structures. Poly-XTAL operations is named after the various operations which lead to its functionalities relating to poly-crystalline structures. It is useful for mathematicians, statistical mechanists, material scientists and computational geologiests. Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [Tutorials and Test Cases](https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases) which you can try out while learning to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. Please use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to as questions/clarifications, put forth your ideas, feature requests, documentation requests and bug/issue reports.
+**PXO** stands for Poly-XTAL Operations. It is an open-source, MATLAB codebase for generating, analyzing and exporting complex space partitions such as multi-phase metallic grain structures. Poly-XTAL operations is named after the various operations which lead to its functionalities relating to poly-crystalline structures. It is useful for mathematicians, statistical mechanists, material scientists and computational geologiests. Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [Tutorials and Test Cases](https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases) which you can try out while learning to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. Please use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to as questions/clarifications, put forth your ideas, feature requests, documentation requests and bug/issue reports. Please visit [HERE](https://github.com/SunilAnandatheertha/PXO/wiki/Installation-instructions) to see installation instructions.
 
 Development and all documentations by: [Sunil Anandatheertha](https://sunilanandatheertha.github.io/). Please start with [wiki page](https://github.com/SunilAnandatheertha/PXO/wiki). It will direct you. To use PXO, no installation is needed yet. Just add the folder to MATLAB path and start using. Sponsorer information will be listed here in the reverse chronological order. Please go through contributing.md to see how you may contribute. Contributer credits will be listed in detail here.
 
