### 2021-07-08 b3b1156
Added an automated test case
