+++ mcodes/Build___INPUT_PARAMS___Struct_test1.m
@@ -1,0 +1,153 @@
+function Build___INPUT_PARAMS___Struct_test1()
+%-------------------------------------------------------------
+% STRUCTURE: Include Lattice details: "latticesizei", "latticesizej", 'Triangular', 'Square', 
+                     % No_Nearest_Neighbour, delE calc by 'descrete' or 'contnuous'
+%-------------------------------------------------------------
+global Lattice MC_Param MC_Loop Viz_options CMDL_display File_Fold_Operations
+%\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
+%\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\
+Lattice.type.square     = 1;
+Lattice.type.triangular = 0;
+Lattice.type.hexagonal  = 0;
+if (Lattice.type.square + Lattice.type.triangular + Lattice.type.hexagonal) > 1
+    Lattice.type.square     = 1;
+    Lattice.type.triangular = 0;
+    Lattice.type.hexagonal  = 0;
+end
+Lattice.type.random     = 0;
+%-------------------------------------------------------------
+Lattice.size.i          = 50;
+Lattice.size.j          = 50;
+Lattice.size.i_incr     = 1.0;
+Lattice.size.j_incr     = 1.0;
+Lattice.size.xmin       = -floor(Lattice.size.i)/2+1;
+Lattice.size.xmax       = +floor(Lattice.size.i)/2;
+Lattice.size.ymin       = -floor(Lattice.size.j)/2+1;
+Lattice.size.ymax       = +floor(Lattice.size.j)/2;
+Lattice.size.xlength    = abs(Lattice.size.xmin)+abs(Lattice.size.xmax);
+Lattice.size.ylength    = abs(Lattice.size.ymin)+abs(Lattice.size.ymax);
+Lattice.size.Pert_Fac   = 0; % Perturbation factor
+[x, y]                  = meshgrid(Lattice.size.xmin : Lattice.size.i_incr : Lattice.size.xmax,...
+                                   Lattice.size.ymin : Lattice.size.i_incr : Lattice.size.ymax);
+Lattice.size.sz1        = size(x, 1);
+Lattice.size.sz2        = size(x, 2);
+Lattice.size.x          = x + Lattice.size.Pert_Fac*rand(Lattice.size.sz1, Lattice.size.sz2);
+Lattice.size.y          = y + Lattice.size.Pert_Fac*rand(Lattice.size.sz1, Lattice.size.sz2);
+%-------------------------------------------------------------
+
+Lattice.q               = 30;
+
+%-------------------------------------------------------------
+Lattice.zener.slsp.want_slsp = 1;
+Lattice.zener.slsp.Vol_Frac  = 5.000; % IN PERCENTAGE
+%-------------------------------------------------------------
+Lattice.zener.slspc.want_slspc            = 0;      % 0: no clustering is not desired
+Lattice.zener.slspc.typeofcluster         = {'circular'   , 1,...
+                                             'oval'       , 0,...
+                                             'rectangluar', 0,...
+                                             'randomized'};
+Lattice.zener.slspc.noslspclusters        = 10;      % Set the number of circular clusters
+Lattice.zener.slspc.clusterdistr          = {'circular'   , 0,...
+                                             'oval'       , 0,...
+                                             'rectangluar', 0,...
+                                             'triangular' , 0,...
+                                             'hexagfonal' , 0,...
[…]
