### 2021-07-08 e45163a
cmd display removed
