### 2021-07-08 #3190 @SunilAnandatheertha
@ksyang2013

Thank you for the review. I have updated the codes and documentation accordingly.

Installation instructions have been detailed in the file “installation_instructions.txt” available in the main repository root directory. A package manager will be developed in the future. It is also provided in the page: https://github.com/SunilAnandatheertha/PXO/wiki/Installation-instructions.

Automated testing can be carried out as outlined in [https://github.com/SunilAnandatheertha/PXO/wiki/Automated-test---1].

It appears the problem “matlab.graphics.internal.name (line 112)” may be because, MATLAB is not reading a full directory on the computer.  The repository has only been tested on Windows 10, so this could be a operating system related issue. Though I am incorporating operating system checking, it is planned to be integrated into PXO distribution in future releases. Could you please copy the repository into a major directory such as “C:\PXO” or “D:\PXO” and run the automated test.

Sincerely
Sunil
