### 2021-07-09 #3190 @SunilAnandatheertha
@Kevin-Mattheus-Moerman  Hi Kevin. I would like to make a few grammatical changes to the paper. Also, over time, I have updated the links to the repository wiki page, and these links need to be updated in the paper. The paper has "a free MATLAB codebase". I would like to remove the word "free", as being open source already implies it. Am I permitted to do these changes. Thank you.

Sincerley
Sunil
