### 2021-07-11 #3190 @Kevin-Mattheus-Moerman
@SunilAnandatheertha yes you can make these changes. You can update the paper in your repository as you see fit and then call `@whedon generate pdf` here to update it. Once you've updated the title I can amend it for this review issue too.
