+++ paper.bib
@@ -9,6 +9,17 @@ volume = {82},
 year = {2015}
 }
 
+@article{RHielscher2015,
+author = {Hielscher, R. and Schaeben, H.},
+doi = {10.1107/S0021889808030112},
+journal = {Journal of Applied Crystallography},
+pages = {1024--1037},
+title = {{A novel pole figure inversion method: specification of the MTEX algorithm}},
+volume = {41},
+part = {6},
+year = {2008}
+}
+
 @book{JMarkKlien1999,
 author = {Mark Klien, J.},
 editor = {Sack, J.-R. and Urrita, J.},
