+++ paper.md
@@ -1,5 +1,5 @@
 ﻿---
-title: 'PXO (Poly-XTAL Operations): MATLAB codebase to generate and analyse complex 2D grain structures '
+title: 'PXO (Poly-XTAL Operations): MATLAB Codebase to Generate, Analyse and Export Complex 2D Spatio-Temporally Gradient Grain Structures'
 tags:
   - grain structure generation
   - texture and grain structure
@@ -15,32 +15,31 @@ bibliography: paper.bib
 
 ---
 # Statement of need
-Mathematicians, statistical mechanists and computational materials scientists are interested in studying the spatiotemporal evolutionary aspects of multi-phased partitioning of an n-dimensional space. Four examples for this are given and the statement of need is provided in each.
+Mathematicians, statistical mechanists and computational materials scientists are interested in studying the spatiotemporal evolutionary aspects of multi-classed partitioning of an n-dimensional space. Four examples for this are given and the statement of need is provided in each.
 
-The 1st example is from mathematics. Here, researchers are interested in the chaotic partitioning of an n-D bounded spatial domain and its spatiotemporal evolution under some governing rules [@JMarkKlien1999]. Mathematicians need a computational platform to study n-dimensional partitions of space.
+The 1st example is from mathematics. Here, researchers are interested in the chaotic partitioning of an n-D bounded spatial domain and its spatiotemporal evolution under some governing rules [@JMarkKlien1999]. A computational platform to study n-dimensional partitions of space is needed to enable this.
 
-The 2nd example is from statistical mechanics. Here researchers are interested in importance sampling Monte-Carlo techniques studying the spatiotemporal evolution of the kinetics and thermodynamics of the distribution of multi-phase systems over some lattice. Exact solutions have been developed for such simple models involving 2 states [@Baxter1989], but for more complex models like the Q-state Pott’s model, an exact model is impractical due to the vastness of the solution space. Here, a computational platform is needed to simulate such models.
+The 2nd example is from statistical mechanics. Here researchers are interested in importance sampling Monte-Carlo techniques studying the spatiotemporal evolution of the kinetics and thermodynamics of the distribution of multi-phase systems over some lattice. Though, exact solutions have been developed for Ising model [@Baxter1989], it is however, impractical for more complex models like the Q-state Pott’s model, owing to the vastness of solution space. Here, a computational platform is needed to simulate such complex models which enables studying their complex behaviours.
 
-The 3rd example is from fundamental computational materials science. Here researchers are interested in grain growth [@Weaire1996], where the temporal evolution of spatial and thermodynamical parameters of multi-phase grain structures [@Anderson1984], [@Anderson1989] is studied at micro- and macro- scales. Here, a computational platform is used to simulate grain growth. A part of this research also touches upon understanding the kinematic and kinetic behaviour of insoluble 2nd phase particles in grain structures [@Srolovitz1984] and how they interact with the grain boundaries. Some of these studies have tried to validate empirical models of grain structure geometry such as the Zener equation [@Manohar1998]. As the shape of the particles influence the Zener drag working against grain boundary evolution during grain growth [@Li1990], and that nature presents irregularly shaped particles, computer models which can consider such particle shape and their spatial distribution becomes very essential. A platform is needed where such studies can be made numerically.
[…]
