### 2021-07-12 #3190 @littlelazy6
@SunilAnandatheertha  I have read the updated paper. Regarding lines 47 to 53, a detailed discussion on VT/ EBSD based CPFEM and the relative capabilities of software like DREAM3D and Neper is included in this work, for your interest https://doi.org/10.1177/14644207211010836.

@Kevin-Mattheus-Moerman After recent documentation updates, I have confirmed the functions of PXO, and think this is a useful and handy tool for microstructural evolution and it fills the gap between other FE simulations tools.  I recommend this work to be published in JOSS.
