+++ paper.md
@@ -23,7 +23,7 @@ The 2nd example is from statistical mechanics. Here researchers are interested i
 
 The 3rd example is from fundamental computational materials science, where researchers are interested in the growth of grains [@Weaire1996]. Here the temporal evolution of spatial and thermodynamical parameters of multi-phase grain structures [@Anderson1984], [@Anderson1989] is studied at micro- and macro- scales. Though phase-field approaches are used, they are however, computational very intensive. Here, a computational platform is used to simulate grain growth over a simplified manner, and Monte-Carlo simulation techniques have been used to address this need [@Anderson1984]. A part of this research also touches upon understanding the kinematic and kinetic behaviour of grain structures in the presence of insoluble 2nd phase particles [@Srolovitz1984] and how they interact with the grain boundaries. Some of these studies have tried to validate emperical models of grain growth (for parameter such as grain growth exponent), and empirical models pertaining to grain structure geometry such as the Zener equation [@Manohar1998]. As the shape of the particles influences the Zener drag working against grain boundary evolution during grain growth [@Li1990], and also that, nature presents irregularly shaped particles, computer models which can consider such particle shape and their spatial distribution becomes very essential in computer simulations of grain growth. In addition to these, it is also important to understand the grain growth kinetics. A software platform is needed where such studies can be made.
 
-The 4th example is from applied computational materials science. Here, researchers use techniques such as crystal plasticity based finite element analysis (CPFEA) to study material’s phase-partitioned thermo-mechanical response and texture evolution under applied thermo-mechanical loads [@Roters2010]. These are usually done at micro- and macro- scales. Computational materials scientists studying multi-scale thermo-mechanical and texture behaviour of poly-crystalline materials need parametric and realistically tessellated geometric morphologies of the constituent phases reflected in the virtual grain structure. The virtual grain structures could be realized using experimentally derived grain structures or by developing a purely computer generated grain structure. Such grain structures are generally used in CPFEA. Though Voronoi tessellated geometries of grain structures have been used before, they are simplifications and do not accurately represent the geometric irregularities presented by nature. However, it is difficult to generate parametric spatially gradient physically realistic grain structures with multiple temporally-low-gradient slices. It is also difficult for many to acquire grain structures using experimental techniques like EBSD and even difficult to process them and use them numerically. Similar needs are also believed to be present in computational geology. A computer software program capable of preparing artificial and physically realistic grain structures for use in FEA will be valuable.
[…]
