+++ paper.bib
@@ -20,6 +20,14 @@ part = {6},
 year = {2008}
 }
 
+@article{YuhuiTu2021,
+author = {Yuhui Tu and Seán B Leen and Noel M Harrison},
+title ={A high-fidelity crystal-plasticity finite element methodology for low-cycle fatigue using automatic electron backscatter diffraction scan conversion: Application to hot-rolled cobalt–chromium alloy},
+journal = {Proceedings of the Institution of Mechanical Engineers, Part L: Journal of Materials: Design and Applications},
+year = {May 2021},
+doi = {10.1177/14644207211010836},
+}
+
 @book{JMarkKlien1999,
 author = {Mark Klien, J.},
 editor = {Sack, J.-R. and Urrita, J.},
