+++ README.md
@@ -1,4 +1,9 @@
 ***
+[![Last commit](https://img.shields.io/github/last-commit/SunilAnandatheertha/PXO)]()
+[![Total downloads](https://img.shields.io/github/downloads/SunilAnandatheertha/PXO/total)]()
+[![Size](https://img.shields.io/github/repo-size/SunilAnandatheertha/PXO)]()
+[![NumCodeLines](https://img.shields.io/tokei/lines/github/SunilAnandatheertha/PXO)]()
+
 ### Introduction
 **PXO** stands for Poly-XTAL Operations. It is an open-source, MATLAB codebase for generating, analyzing and exporting complex space partitions such as multi-phase metallic grain structures. Poly-XTAL operations is named after the various operations which lead to its functionalities relating to poly-crystalline structures. It is useful for mathematicians, statistical mechanists, material scientists and computational geologiests. Read the [MOTIVATION](https://github.com/SunilAnandatheertha/PXO/wiki/Motivation) for developing PXO, along with a set of [VALIDATION](https://github.com/SunilAnandatheertha/PXO/wiki/Validation) results, current [LIMITATIONS](https://github.com/SunilAnandatheertha/PXO/wiki/Limitations) of PXO, [Tutorials and Test Cases](https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases) which you can try out while learning to use. In case you use MTEX and would like to expand your research scope by being able to generate your own custom grain structures and crystallographic textures, you can use PXO to do so. Please use the [DISCUSSION](https://github.com/SunilAnandatheertha/PXO/discussions) page to as questions/clarifications, put forth your ideas, feature requests, documentation requests and bug/issue reports. Please visit [HERE](https://github.com/SunilAnandatheertha/PXO/wiki/Installation-instructions) to see installation instructions.
 
@@ -11,9 +16,3 @@ Please cite PXO as: Anandatheertha, S., (2021). PXO (Poly-XTAL Operations): MATL
 ## License
 Refer licensing file for more information.
 
-***
-
-[![Last commit](https://img.shields.io/github/last-commit/SunilAnandatheertha/PXO)]()
-[![Total downloads](https://img.shields.io/github/downloads/SunilAnandatheertha/PXO/total)]()
-[![Size](https://img.shields.io/github/repo-size/SunilAnandatheertha/PXO)]()
-[![NumCodeLines](https://img.shields.io/tokei/lines/github/SunilAnandatheertha/PXO)]() 
