### 2021-07-20 #3190 @Kevin-Mattheus-Moerman
@SunilAnandatheertha looks like we are ready for the final steps.

I have just read your paper and provide the following required minor changes:
- [ ] Check `...developed for Ising model...`, perhaps this should read `...developed for the Ising model...`
- [ ] Check `... is used to simulate grain growth over a simplified manner, and...`, should  this read  `... is used to simulate grain growth in a simplified manner, and...`?
- [ ] Check `emperical` this should read `empirical`
- [ ] Check `softwares` this should read `software`
- [ ] Check `relavant` this should read `relevant`
- [ ] Check `powerfull` this should read `powerful`
- [ ] Check the use of the acronym VT for Voronoi. I propose you either remove it and not use it, or use it but introduce it when Voronoi is first mentioned and then use it after (currently Voronoi is used several times in full before the acronym is used). I recommend minimising acronym use where possible but I leave this up to you.
- [ ] The same as above about the acronym for Monte-Carlo simulations (MCS). Again I recommend removing it but if used use it consistently.
- [ ] Check `...to study material’s phase-transitioned...`, perhaps this should read  `...to study a material’s phase-transitioned...`
- [ ] Check `Computational materials scientists...`, should this read `Computational material scientists...`
- [ ] The following sounds like it needs to be rephrased `...parametric spatially gradient physically realistic grain structures...`, perhaps this sounds clearer/better: `...parametric representations and spatial gradients of physically realistic grain structures...`. Please adjust as you see fit.
- [ ] Consider rephrasing `Similar needs are also believed to be present in computational geology.`, in particular to avoid the use of the word `believed. Perhaps something like the following would be better?: `Furthermore, such methods are also relevant to computational geology. `
- [ ] The acronym `FEA` is not explained please fix or avoid its use. Again it is also used in full as `finite element analysis`, after its first use so please be consistent again or remove.
- [ ] Check the use of PXO and  ‘PXO’, please choose one style and use it throughout. Furthermore, it looks like you are trying to use it as `PXO`, but perhaps you are using the wrong style of inverted commas, if you want this style use
`PXO`
```
- [ ] Check `...generated grain structure...`, it looks like this should read `...generated grain structures...`.
- [ ] Similarly add an `s` to `...of VT grain structure.` so it reads `...of VT grain structures.`.
- [ ] `documentations` -> `documentation`
- [ ] `easy to use` -> `easy-to-use`
- [ ] `...ample documentation and growing scientific ...` -> `...ample documentation, and a growing scientific ...`
- [ ] `...As PXO offers tool to export ...` -> `...As PXO offers a tool to export ...`
- [ ] `...complex grain87structure distribution.` ->  `...complex grain87structure distributions.`
- [ ] `Above figure shows...` -> `The above figure shows...`
- [ ] `...to help the user in getting started.` -> `...to help users to get started.`
- [ ] `... resource offered by the institutions ...` -> `...resources offered by the following institutions: ...`
- [ ] In relation to ` PESIT (now called as PES University)` can you simply use just the latest name?
- [ ] Your paper is about to be accepted in JOSS. We recommend you thoroughly proofread the paper yourself once more at this point and also verify the author names and affiliations are accurate, as well as the acknowledgements. I also did not comment to much on the use of commas, however some of the longer sentences can likely benefit from this so review these also.

Once you have processed the above please update the paper here by calling `@whedon generate pdf`, next please complete the following:
- [ ] Can you archive a copy of the software on ZENODO and report back here with the DOI or the archived version? You can do this manually or you can follow these steps: https://guides.github.com/activities/citable-code/
- [ ] Once the archive is created check that the archive meta data, like the **title**, and the **authors**, matches those of the paper. You may need to manually edit this information. You can also add ORCID profile links for all authors at this point if not present already.
- [ ] Can you confirm what the published version tag should be? Is it still at `10.0.2`?
