+++ paper.md
@@ -15,31 +15,31 @@ bibliography: paper.bib
 
 ---
 # Statement of need
-Mathematicians, statistical mechanists and computational materials scientists are interested in studying the spatiotemporal evolutionary aspects of multi-classed partitioning of an n-dimensional space. Four examples for this are given and the statement of need is provided in each.
+Mathematicians, statistical mechanists and computational material scientists are interested in studying the spatiotemporal evolutionary aspects of multi-classed partitioning of an n-dimensional space. Four examples for this are given and the statement of need is provided in each.
 
 The 1st example is from mathematics. Here, researchers are interested in the chaotic partitioning of an n-D bounded spatial domain and its spatiotemporal evolution under some governing rules [@JMarkKlien1999]. A computational platform to study n-dimensional partitions of space is needed to enable this.
 
-The 2nd example is from statistical mechanics. Here researchers are interested in importance sampling Monte-Carlo techniques studying the spatiotemporal evolution of the kinetics and thermodynamics of the distribution of multi-phase systems over some lattice. Though, exact solutions have been developed for Ising model [@Baxter1989], it is however, impractical for more complex models like the Q-state Pott’s model, owing to the vastness of solution space. Here, a computational platform is needed to simulate such complex models which enables studying their complex behaviours.
+The 2nd example is from statistical mechanics. Here researchers are interested in importance sampling Monte-Carlo techniques studying the spatiotemporal evolution of the kinetics and thermodynamics of the distribution of multi-phase systems over some lattice. Though, exact solutions have been developed for the Ising model [@Baxter1989], it is however, impractical for more complex models like the Q-state Pott’s model, owing to the vastness of solution space. Here, a computational platform is needed to simulate such complex models which enables studying their complex behaviours.
 
-The 3rd example is from fundamental computational materials science, where researchers are interested in the growth of grains [@Weaire1996]. Here the temporal evolution of spatial and thermodynamical parameters of multi-phase grain structures [@Anderson1984], [@Anderson1989] is studied at micro- and macro- scales. Though phase-field approaches are used, they are however, computational very intensive. Here, a computational platform is used to simulate grain growth over a simplified manner, and Monte-Carlo simulation techniques have been used to address this need [@Anderson1984]. A part of this research also touches upon understanding the kinematic and kinetic behaviour of grain structures in the presence of insoluble 2nd phase particles [@Srolovitz1984] and how they interact with the grain boundaries. Some of these studies have tried to validate emperical models of grain growth (for parameter such as grain growth exponent), and empirical models pertaining to grain structure geometry such as the Zener equation [@Manohar1998]. As the shape of the particles influences the Zener drag working against grain boundary evolution during grain growth [@Li1990], and also that, nature presents irregularly shaped particles, computer models which can consider such particle shape and their spatial distribution becomes very essential in computer simulations of grain growth. In addition to these, it is also important to understand the grain growth kinetics. A software platform is needed where such studies can be made.
[…]
