+++ paper.bib
@@ -20,6 +20,19 @@ part = {6},
 year = {2008}
 }
 
+@article{MTEX2Gmsh2020,
+  doi = {10.21105/joss.02094},
+  url = {https://doi.org/10.21105/joss.02094},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {52},
+  pages = {2094},
+  author = {Dorian Depriester and R\'egis Kubler},
+  title = {MTEX2Gmsh: a tool for generating 2D meshes from EBSD data},
+  journal = {Journal of Open Source Software}
+}
+
 @article{YuhuiTu2021,
 author = {Yuhui Tu and Seán B Leen and Noel M Harrison},
 title ={A high-fidelity crystal-plasticity finite element methodology for low-cycle fatigue using automatic electron backscatter diffraction scan conversion: Application to hot-rolled cobalt–chromium alloy},
