### 2021-07-20 83992de
Deleting the older version of the paper
