### 2021-07-20 f6d4f8b
Deleting the older version of the paper
