+++ mcodes/GeometryWarper.m
@@ -1,7 +1,0 @@
-function [varargout] = AtomicGeometryWarper(varargin)
-
-switch varargin{1}
-    case 'AGG'
-end
-
-end
