+++ mcodes/SDMantra.m
@@ -1,15 +1,0 @@
-function SDMantra(varargin)
-
-switch varargin{1}
-    case 'saraswati'
-        SCMD('seperator','type03',5,1)
-        fprintf('saraswati namasthubhyam varade kAmarUpiNi\nvidhyAraMbham karishyAmi siddhirbhavathu me sadA\n')
-        SCMD('seperator','type02',5,1)
-        disp('may dEvi saraswati bless you with awesome programming skills..')
-        SCMD('seperator','type02',5,1)
-        fprintf('take care of your thoughts when you are alone\ntake care of your words when you are with people\n')
-        SCMD('seperator','type04',5,1)
-        pause(0.25)
-end
-
-end
