### 2021-07-20 #3190 @Kevin-Mattheus-Moerman
@SunilAnandatheertha

Thanks for implementing many of those points

- [ ] Check above several boxes remain unticked
- [ ] Please edit the ZENODO archive so that the license matches [your BSD 3-Clause License](https://github.com/SunilAnandatheertha/PXO/blob/main/LICENSE.md).
- [ ] On the ZENODO archive please remove the `Rights holder(s)` field
- [ ] The version tag should be one of the following: https://github.com/SunilAnandatheertha/PXO/tags so I take it you mean `V10.1.0`.
