+++ mcodes/results/datafiles/statematrices/s100mcs1.txt
@@ -1,50 +1,0 @@
-6	6	6	9	9	9	7	7	8	7	7	7	7	7	8	8	8	8	8	6	6	6	6	6	6	1	1	1	1	1	1	2	2	2	7	7	7	7	7	7	7	7	7	10	10	10	10	10	10	10
-6	6	6	9	9	9	9	8	8	7	7	7	7	7	8	8	8	8	8	8	6	6	6	6	6	1	1	1	1	1	1	1	7	7	7	7	7	7	7	7	7	7	2	10	10	10	10	10	10	6
-6	6	6	9	9	9	10	8	8	8	7	7	7	7	8	8	8	8	8	8	5	5	5	5	5	1	1	1	1	1	1	7	7	7	7	7	7	7	7	7	7	7	10	10	10	10	10	10	10	10
-6	6	2	9	9	10	10	10	8	1	1	1	1	1	8	8	8	8	8	5	5	5	5	5	9	1	9	1	1	1	1	7	7	7	7	7	7	7	7	7	7	7	10	10	10	10	10	10	10	10
-2	2	2	2	10	10	10	10	10	1	1	1	1	4	4	4	4	4	5	5	5	5	5	5	5	9	9	9	9	9	7	7	7	7	7	7	7	7	7	7	4	4	4	10	10	10	10	10	10	10
-2	2	2	2	10	10	10	10	10	1	1	1	1	4	4	4	4	4	5	5	5	5	5	5	5	9	9	9	9	9	9	7	7	7	7	7	7	7	7	4	4	4	4	4	10	10	10	10	10	10
-2	2	2	2	10	10	10	10	10	1	1	1	1	4	4	4	4	4	5	5	5	5	5	5	5	9	9	9	9	9	9	7	7	7	7	7	7	7	4	4	4	4	4	4	4	10	10	10	10	10
-2	2	2	2	2	10	10	10	7	7	1	1	1	4	4	4	4	3	3	5	5	5	5	5	5	9	9	9	9	9	9	9	7	7	7	7	7	4	4	4	4	4	4	4	4	4	10	10	10	5
-5	2	2	2	2	10	10	7	7	7	7	7	4	4	4	4	3	3	3	3	5	5	5	5	9	9	9	9	9	9	9	9	9	7	7	7	7	4	4	4	4	4	4	4	4	4	10	5	5	5
-5	5	2	2	2	2	7	7	7	7	7	7	7	4	3	3	3	3	3	3	6	6	6	6	9	9	9	9	9	9	9	9	10	10	7	7	7	7	4	4	4	4	4	4	4	4	5	5	5	5
-5	5	5	2	2	2	7	7	7	7	7	7	6	3	3	3	3	3	3	3	6	6	6	6	6	4	4	9	9	9	9	10	10	10	10	10	7	4	4	4	4	9	4	4	4	5	5	5	5	5
-5	5	5	5	8	8	8	7	7	7	7	7	3	3	3	3	3	3	3	6	6	6	6	6	4	4	4	4	9	9	10	10	10	10	10	10	10	4	4	4	4	4	4	4	4	5	5	5	5	5
-5	5	5	8	8	8	8	8	8	8	8	3	3	3	3	3	3	3	6	6	6	6	6	4	4	4	4	4	4	10	10	10	10	10	10	10	10	10	4	4	4	4	4	4	5	5	5	5	5	5
-5	5	8	8	8	8	8	8	8	8	8	8	3	3	3	3	3	6	6	6	6	6	4	4	4	4	4	4	4	10	10	10	10	10	10	10	10	10	4	4	4	4	4	4	5	5	5	5	5	5
[…]
