+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_4_incr =7.9305.txt
@@ -1,1 +1,0 @@
-7.9305
