+++ mcodes/results/datafiles/statematrices/s2000mcs1.txt
@@ -1,50 +1,0 @@
-3	3	3	3	3	3	3	3	3	3	3	3	3	4	4	4	4	4	4	4	4	4	4	18	18	16	16	16	16	16	16	16	16	30	30	27	27	27	27	27	27	27	27	27	27	27	27	27	27	27
-3	3	3	3	3	3	3	3	3	3	3	3	3	4	4	4	4	4	4	4	4	4	4	4	16	16	16	16	16	16	16	16	16	16	16	27	27	27	27	27	27	27	27	27	27	27	27	27	27	3
-3	3	3	3	3	21	3	3	3	3	3	3	3	6	4	4	4	4	4	4	4	4	4	4	16	16	16	16	16	16	16	16	16	16	26	26	27	27	27	27	27	27	27	27	27	27	27	27	27	3
-3	3	3	3	3	3	3	3	3	3	3	3	3	3	4	4	4	4	4	4	4	4	4	17	16	16	16	16	16	16	16	16	16	26	26	26	26	27	27	27	27	27	27	27	27	27	30	23	23	3
-3	3	3	3	3	3	3	3	3	3	3	3	3	18	4	4	4	4	4	4	4	4	4	16	16	16	16	16	16	16	16	16	16	26	26	26	26	26	27	27	27	27	27	27	27	23	23	23	23	23
-23	3	3	3	3	3	3	3	3	3	3	3	3	7	4	4	4	4	4	4	4	4	4	16	16	16	16	16	16	16	16	16	16	26	26	26	26	26	26	27	27	27	27	27	23	23	23	23	23	23
-23	3	3	3	3	3	3	3	3	3	3	3	22	22	22	22	22	4	4	4	4	4	16	16	16	16	16	16	16	16	16	4	4	4	26	26	26	26	2	2	2	2	27	23	23	23	23	23	23	23
-23	3	3	3	3	3	3	3	3	3	3	22	22	22	22	22	22	22	22	4	4	17	17	16	16	16	16	16	16	4	4	4	4	4	4	4	26	2	2	2	2	2	2	23	23	23	23	23	23	23
-23	23	3	3	3	3	3	3	3	3	22	22	22	22	22	22	22	22	22	22	17	17	17	17	16	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	23	23	23	23	23	23
-23	23	23	3	3	3	3	3	3	22	22	22	22	22	22	22	22	22	22	22	22	17	17	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	23	23	23	23	23
-23	23	23	9	9	3	3	3	22	22	22	22	22	22	22	22	22	22	22	22	22	22	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	23	23	23	23
-23	23	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	22	22	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	23	23	23
-23	9	9	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	22	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	2	2	7	7	7
-7	9	9	9	9	9	9	9	9	9	22	22	22	22	22	22	22	22	22	22	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	4	2	2	2	2	2	2	2	2	7	7	7	7
[…]
