+++ mcodes/results/datafiles/graindata/grainsizes/averagegrainsize_using11_HORIZONTAL_intercepts.txt
@@ -1,4 +1,0 @@
-5.0376
-6.5549
-7.3744
-7.9343
