+++ mcodes/results/datafiles/graindata/grainsizes/A_A_AGSvsMCS_M_intincr_1_and_2 =7.9941.txt
@@ -1,1 +1,0 @@
-7.9941
