+++ mcodes/results/datafiles/graindata/Qgrainelements/MCS500/1.txt
@@ -1,73 +1,0 @@
-769
-770
-818
-819
-820
-821
-822
-823
-867
-868
-869
-870
-871
-872
-873
-874
-875
-916
-917
-918
-919
-920
-921
-922
-923
-924
-925
-965
-966
-967
-968
-969
-970
-971
-972
-973
-974
-975
-1016
-1017
-1018
-1019
-1020
-1021
-1022
-1023
-1024
-1025
-1066
-1067
-1068
-1069
-1070
-1071
-1072
-1073
-1074
-1075
-1119
-1120
-1121
-1122
-1123
-1124
-1169
-1170
-1171
-1172
-1173
-1176
-1220
-1221
-1584

[…]
