+++ mcodes/pradyumna.m
@@ -1,115 +1,0 @@
-close all;clear all; hold on
-NOfPointsOnFitCurve = 60;
-xp150_750x750_icr00625_icvf        = [ 00.0200 00.0300 00.0500 00.0750 00.1000 00.2500 00.5000 00.7500 01.0000 ];
-xp150_750x750_icr00625_icvf_ags_T1 = [ 85.4489 72.2753 64.3134 56.3070 51.5780 44.5293 40.3929 40.4486 51.5036 ];
-xp150_750x750_icr00625_icvf_ags_T2 = [ 81.3436 73.2063 64.3134 59.6895 52.9240 42.1042 40.3000 43.1652 51.1691 ];
-xp150_750x750_icr00625_icvf_ags = (xp150_750x750_icr00625_icvf_ags_T1 + xp150_750x750_icr00625_icvf_ags_T2)/2;
-xp150_750x750_icr00625_icvf_ags = xp150_750x750_icr00625_icvf_ags/max(xp150_750x750_icr00625_icvf_ags);
-x00625 = xp150_750x750_icr00625_icvf;
-y00625 = xp150_750x750_icr00625_icvf_ags;
-z00625 = 0.00625*ones(size(xp150_750x750_icr00625_icvf_ags));
-hold on; view(30,30)
-plot3(x00625,y00625,z00625, 'or', 'LineWidth', 1, 'MarkerFaceColor', 'r', 'MarkerSize', 4)
-[xData00625, yData00625] = prepareCurveData( x00625, y00625 );
-ft   = fittype( 'rat31' ); opts = fitoptions( ft ); opts.Display = 'Off';
-opts.Lower       = [-Inf -Inf -Inf -Inf -Inf];
-opts.StartPoint  = [0.02 0.02 0.02 0.02 0.02];
-opts.Upper       = [Inf Inf Inf Inf Inf];
-[fitresult00625, gof] = fit( xData00625, yData00625, ft, opts );
-xstart           = min(x00625);
-xend             = max(x00625);
-xfitdata00625 = linspace(xstart, xend, NOfPointsOnFitCurve);
-yfitdata00625 = ( (fitresult00625.p1 * xfitdata00625.^3) + (fitresult00625.p2 * xfitdata00625.^2) + (fitresult00625.p3 * xfitdata00625) + (fitresult00625.p4) )./ (xfitdata00625 + fitresult00625.q1);
-zfitdata00625 = 0.00625*ones(size(xfitdata00625));
-plot3(xfitdata00625, yfitdata00625, zfitdata00625,'r','LineWidth',2)
-xmin(1) = xfitdata00625(yfitdata00625 == min(yfitdata00625));
-ymin(1) = yfitdata00625(yfitdata00625 == min(yfitdata00625));
-zmin(1) = zfitdata00625(yfitdata00625 == min(yfitdata00625));
-
-
-
-
-
-xp150_750x750_icr01250_icvf = [ 0.020 0.025 0.030 0.035 0.040 0.050 0.075 0.100 0.250 0.500 0.750 1.000 ];
-xp150_750x750_icr01250_icvf_ags_T1 = [ 51.2031 47.0616 46.6017 47.3751 41.6098 39.2802 33.7059 29.7026 22.8319 21.7933 22.0248 42.8334 ];
-xp150_750x750_icr01250_icvf_ags_T2 = [ 55.4995 49.3905 46.6426 47.3751 41.5526 37.9948 34.8839 30.9316 22.8319 21.3443 22.0248 42.7954 ];
-xp150_750x750_icr01250_icvf_ags    = (xp150_750x750_icr01250_icvf_ags_T1 + xp150_750x750_icr01250_icvf_ags_T2)/2;
[…]
