+++ mcodes/trial1.m
@@ -1,144 +1,0 @@
-% this is a trial fea code for LINEAR springs
-% EleCon = [elem1 nodeiOfelem1 nodejOfelem1;
-%           elem2 nodeiOfelem2 nodejOfelem2;];
-% ExEleCon = [elem1 Kelem1 nodeiOfelem1 nodejOfelem1;
-%             elem2 Kelem2 nodeiOfelem2 nodejOfelem2;];
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%% ELEMENT INFORMATION
-EleCon = [1 1 2;
-          2 2 3;
-          3 3 4;
-          4 4 5;
-          5 5 6;
-          6 6 7;
-          7 7 8;
-          8 8 9;
-          9 9 10;
-          10 10 11]; % EleCon - Element Connectivity
-k      = [100*rand;
-          100*rand;
-          100*rand
-          100*rand
-          100*rand
-          100*rand
-          100*rand
-          100*rand
-          100*rand
-          100*rand]; % k values in kN/m
-NoOfElements   = size(EleCon,1);
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%% BOUNDARY CONDITIONS ---- LOADS:
-%   there are no loads applied on nodes 2 to (nd_sys - 1)
-%   Hence, if nd_sys = 8, in Kd = F, F values for nodes 2,3,4..,7 are zero
-%   P is applied on nd.no. 8
-%   But, ther exists a reaction force on node 1 as it is constrained against any motion
-NodalForce = [NaN;
-              0;
-              0;
-              0;
-              -10;
-              0;
-              0;
-              0;
-              0;
-              0;
-              15;];
-% units: kN
-% NaN entries marks unkown force values. These are to be calculated
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%% BOUNDARY CONDITIONS ---- DISPLACEMENTS:
-%   only node 1 is constrained to move. 
-%   hence u1 = 0
-NodalDisp = [0;
-             NaN;
-             NaN;
-             NaN;
-             NaN;
-             NaN;
-             NaN;
-             NaN;
-             NaN;
-             NaN;
-             NaN;];
-% units: m
-% NaN entries marks unkown displacement values. These are to be calculated
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%% FORMULATE ELEMENT STIFFNESS MATRICES, AND INCLUDE IT AND ELEMENT NODE NUMBERS, IN KLCL_cell
-KLCL_cell = cell(NoOfElements,2);
-for elem = 1:NoOfElements
-    KLCL_cell{elem,1} = [EleCon(elem,2) EleCon(elem,3)];
-    KLCL_cell{elem,2} = k(elem)*[1 -1; -1  1];
-end
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%% CALCULATE KGBL{1,1}
-nodes_in_sys = unique(reshape(EleCon(:,2:3),1,numel(EleCon(:,2:3)))); % these are the nodes in the FE system
-NoOfNodes    = numel(nodes_in_sys); % number of nodes in the FE system
-MinNdNum_sys = min(nodes_in_sys);   % minimum node number in the FE system
-MaxNdNum_sys = max(nodes_in_sys);   % maximum node number in the FE system
[…]
