+++ mcodes/tryingThis.m
@@ -1,1 +1,0 @@
-[o1, o2, o3] = HiThere(2, 3)
