### 2021-07-28 c836fac
Delete unwanted files and also test generated data
