+++ mcodes/simparameters/initialfinalmcs.txt
@@ -1,2 +1,2 @@
 1
-100
+1000
