+++ mcodes/simparameters/simpar.txt
@@ -1,2 +1,2 @@
 10
-100
+1000
