+++ mcodes/results/colormatrix.txt
@@ -1,10 +1,10 @@
-0.82911	0.84829	0.23073
-0.73298	0.15117	0.40413
-0.36253	0.58585	0.7406
-0.99035	0.80131	0.0034284
-0.76143	0.15534	0.95716
-0.10108	0.15286	0.61593
-0.23972	0.86705	0.18572
-0.27225	0.28858	0.51468
-0.055229	0.89345	0.72977
-0.046697	0.32331	0.44833
+0.71757	0.57762	0.69102
+0.89917	0.48122	0.72288
+0.15974	0.27136	0.031959
+0.13039	0.90279	0.79048
+0.26307	0.37482	0.65862
+0.62074	0.32822	0.15518
+0.97639	0.19777	0.49481
+0.30515	0.39104	0.38272
+0.93176	0.54017	0.50655
+0.39244	0.41326	0.61653
