+++ mcodes/results/datafiles/e/zener_indices.txt
@@ -1,27 +1,27 @@
-244
-780
-817
-1024
-746
-611
-2382
-251
-796
-2042
-1917
-401
-89
-274
-762
-912
-1300
-2040
-2323
+1407
+1475
+1119
+1805
+2120
 1939
-96
-2102
-2486
-2061
-731
+2381
+2137
+252
+103
+54
+15
+497
+1817
+1393
+1978
+1347
+1718
+364
+782
+2321
+212
+128
+544
+941
 -1
 -1
