+++ mcodes/txtfiledet.txt
@@ -1,1 +1,10 @@
 100	2
+200	2
+300	2
+400	2
+500	2
+600	2
+700	2
+800	2
+900	2
+1000	2
