+++ mcodes/Build___INPUT_PARAMS___Struct.m
@@ -104,7 +104,7 @@ MC_Param.StartFrom_MCStep = 1;
 
 
 
-MC_Param.Num_MC_Steps     = 1000; %<<|<<|<<|<<|<<|<<|<<|<<|<<|<<|<<|
+MC_Param.Num_MC_Steps     = 100; %<<|<<|<<|<<|<<|<<|<<|<<|<<|<<|<<|
 
 
 
