### 2021-07-28 #3190 @Kevin-Mattheus-Moerman
@SunilAnandatheertha that is okay, I'll check them when ready. Note they remain unchecked now since those steps appear not to have been changed in the latest paper. I repeat them again below:

- [ ] Check `softwares` this should read `software`
- [ ] The acronym `FEA` is not explained please fix or avoid its use. Again it is also used in full as `finite element analysis`, after its first use so please be consistent again or remove.
- [ ] ...`to help the users to getting started` -> ...`to help users to get started.`
