+++ paper.md
@@ -32,12 +32,12 @@ Poly-XTAL operations (`PXO`) is a MATLAB based computational platform which can 
 All documentation, tutorials and test cases may be accessed from [PXO/Wiki](https://github.com/SunilAnandatheertha/PXO/wiki).
 
 # State of the field
-Over the past few decades, MATLAB has emerged to be an easy-to-use, powerful and robust, high-level computational platform with ample documentation, and a fast growing scientific and engineering community. Poly-XTAL Operations, written in MATLAB, offers the relevant research community, an added advantage over other programming softwares, as MATLAB has a flat to moderate learning curve. Options like MCPM [@JKMason2015], NEPER [@Quey2014] and DREAM3D [@MAGroeber2014] are available to model grain structures [@YuhuiTu2021]. Though, these are powerful environments to deal with grain structure calculations, researchers from a non-software background may find it hard to use and/or modify them to their specific requirements. Implementing user developed solver algorithms in `PXO` is easy. `PXO` offers the added advantage of being able to be linked with `MTEX` [@RHielscher2015] to deal with crystallographic texture calculations and advanced grain structure analysis. As `PXO` offers tools to export as many available temporal slices of grain structures as the user wishes, to a format able to be used in ABAQUS software for FEA (using `MTEX2Gmesh` [@MTEX2Gmsh2020]), it offers a comparatively unique, viable and fast solution to researchers working in the field of Integrated Computational Materials Engineering (ICME) investigating structure-property-behaviour relationships. Additionally, `PXO` is able to combine this advantage with the ability to simulate grain growth in the presence of particles, particle clusters, fibres, whiskers and temperature fields. Another unique advantage is the capability of generating morphologically complex grain structure distributions.
+Over the past few decades, MATLAB has emerged to be an easy-to-use, powerful and robust, high-level computational platform with ample documentation, and a fast growing scientific and engineering community. Poly-XTAL Operations, written in MATLAB, offers the relevant research community, an added advantage over other programming software, as MATLAB has a flat to moderate learning curve. Options like MCPM [@JKMason2015], NEPER [@Quey2014] and DREAM3D [@MAGroeber2014] are available to model grain structures [@YuhuiTu2021]. Though, these are powerful environments to deal with grain structure calculations, researchers from a non-software background may find it hard to use and/or modify them to their specific requirements. Implementing user developed solver algorithms in `PXO` is easy. `PXO` offers the added advantage of being able to be linked with `MTEX` [@RHielscher2015] to deal with crystallographic texture calculations and advanced grain structure analysis. As `PXO` offers tools to export as many available temporal slices of grain structures as the user wishes, to a format able to be used in ABAQUS software for finite element analysis (using `MTEX2Gmesh` [@MTEX2Gmsh2020]), it offers a comparatively unique, viable and fast solution to researchers working in the field of Integrated Computational Materials Engineering (ICME) investigating structure-property-behaviour relationships. Additionally, `PXO` is able to combine this advantage with the ability to simulate grain growth in the presence of particles, particle clusters, fibres, whiskers and temperature fields. Another unique advantage is the capability of generating morphologically complex grain structure distributions.
 
 # Example grain structures
 ![Example grain structures](Paper_images/example1.jpg)
 
-The above figure shows some sample grain structures and space partitions, which can be generated using `PXO`. [Tutorials](https://github.com/SunilAnandatheertha/PXO/wiki/Tutorials-and-test-cases) have been provided to help the users to getting started.
[…]
