### 2021-07-29 #3190 @SunilAnandatheertha
@Kevin-Mattheus-Moerman and the entire JOSS team: JOSS is a very good initiative. Thank you for this platform. Thank you also for considering the work in your journal. I am happy to see the acceptance. @ksyang2013 @yangbai90 @littlelazy6 Thank you so much for your review comments, they have been really helpful. Thank you all.
